Compare Diversity Measures

Compare Diversity Measures

Usage

diversity.compare(
  x,
  group,
  R = 1000,
  base = exp(1),
  na.omit = TRUE,
  global.test = TRUE,
  pairwise.test = TRUE,
  bootstrap.ci = TRUE,
  diversity.profile = TRUE,
  p.adjust.method = c("bonferroni", "holm"),
  ci.conf = 0.95,
  ci.type = c("perc", "bca"),
  q = seq(0, 3, 0.1),
  parallel = c("no", "multicore", "snow"),
  ncpus = 1L,
  cl = NULL
)

Arguments

x

A factor vector of categories (e.g., species, traits). The frequency of each level is treated as the abundance of that category.

group

A factor vector indicating the group of each observation. Must have the same length as x.

R

Integer specifying the number of permutations. Default is 1000.

base

The logarithm base to be used for computation of shannon family of diversity indices. Default is exp(1).

na.omit

logical. If TRUE, missing values (NA) are ignored and not included as a distinct factor level for computation. Default is TRUE.

global.test

logical. If TRUE performs the global permutation tests for the diversity measures. Default is TRUE.

pairwise.test

logical. If TRUE performs the pairwise permutation tests for the diversity measures. Default is TRUE.

bootstrap.ci

logical. If TRUE computes the bootstrap confidence intervals for the diversity measures. Default is TRUE.

diversity.profile

logical. If TRUE diversity profiles. Default is TRUE.

p.adjust.method

(perm.test.pairwise only) Method for adjusting p-values for multiple comparisons. Options include "bonferroni" and "holm". Default is "bonferroni".

ci.conf

Confidence level of the bootstrap interval. Default is 0.95.

ci.type

A vector of character strings representing the type of intervals required. The options are c("perc", "bca").

q

The order of the parametric index.

parallel

The type of parallel operation to be used (if any). If missing, the default is taken from the option "boot.parallel" (and if that is not set, "no").

ncpus

integer: number of processes to be used in parallel operation: typically one would chose this to the number of available CPUs.

cl

An optional parallel or snow cluster for use if parallel = "snow". If not supplied, a cluster on the local machine is created for the duration of the boot call.

Value

A list with the following elements.

Diversity Indices

A data frame of the different diversity indices computed for each group.

Global Test

A data frame of results of global permutation test including the test statistic (weighted sum of squares between group summary indices) and the p value for the different diversity indices.

Pairwise Test

A list of the following data frames.

p-value

A data frame of p values for each between group comparison for different diversity measures.

cld

A data frame of compact letter displays of significant differences among groups for different diversity measures.

Bootstrap CIs

A data frame of lower and upper bootstrap confidence intervals computed for each group in different diversity measures.

Diversity profiles

A list of data frames of Hill, Renyi and Tsallis diversity profiles computed for each group.

Examples

library(EvaluateCore)

pdata <- cassava_CC

qual <- c("CUAL", "LNGS", "PTLC", "DSTA", "LFRT", "LBTEF", "CBTR", "NMLB",
          "ANGB", "CUAL9M", "LVC9M", "TNPR9M", "PL9M", "STRP", "STRC",
          "PSTR")

# Convert qualitative data columns to factor
pdata[, qual] <- lapply(pdata[, qual], as.factor)

str(pdata)
#> 'data.frame':	168 obs. of  26 variables:
#>  $ CUAL  : Factor w/ 4 levels "Dark green","Green purple",..: 3 1 2 2 2 2 4 2 2 1 ...
#>  $ LNGS  : Factor w/ 3 levels "Long","Medium",..: 3 1 2 2 2 2 2 1 1 1 ...
#>  $ PTLC  : Factor w/ 5 levels "Dark green","Green purple",..: 3 4 4 4 4 5 4 2 2 5 ...
#>  $ DSTA  : Factor w/ 5 levels "Absent","Central part",..: 1 5 5 5 5 5 5 4 2 5 ...
#>  $ LFRT  : Factor w/ 4 levels "25-50% leaf retention",..: 1 1 1 1 3 2 2 2 2 2 ...
#>  $ LBTEF : Factor w/ 6 levels "0","1","2","3",..: 3 1 2 1 4 5 4 4 3 2 ...
#>  $ CBTR  : Factor w/ 3 levels "Cream","White",..: 2 2 2 2 1 2 1 1 1 1 ...
#>  $ NMLB  : Factor w/ 9 levels "0","1","2","3",..: 3 1 2 1 4 4 4 3 3 4 ...
#>  $ ANGB  : Factor w/ 4 levels "150-300","450-600",..: 1 4 1 4 2 2 2 1 2 2 ...
#>  $ CUAL9M: Factor w/ 5 levels "Dark green","Green",..: 1 1 3 5 3 3 5 5 5 4 ...
#>  $ LVC9M : Factor w/ 5 levels "Dark green","Green",..: 4 3 3 3 3 1 3 1 4 3 ...
#>  $ TNPR9M: Factor w/ 5 levels "1","2","3","4",..: 5 5 4 2 5 4 2 5 5 5 ...
#>  $ PL9M  : Factor w/ 2 levels "Long (25-30cm)",..: 2 2 1 1 1 1 1 1 2 2 ...
#>  $ STRP  : Factor w/ 4 levels "Absent","Intermediate",..: 2 3 1 1 1 1 4 1 1 4 ...
#>  $ STRC  : Factor w/ 2 levels "Absent","Present": 2 2 1 2 1 1 2 1 1 2 ...
#>  $ PSTR  : Factor w/ 2 levels "Irregular","Tending toward horizontal": 1 2 2 2 1 2 2 2 1 2 ...
#>  $ NMSR  : num  6 2 6 2 20 13 4 14 10 5 ...
#>  $ TTRN  : num  3 0.5 3 2 5 ...
#>  $ TFWSR : num  1.4 2.6 1.2 1.6 5 7 4.2 2.8 2.8 4 ...
#>  $ TTRW  : num  0.7 0.65 0.6 1.6 1.25 ...
#>  $ TFWSS : num  1 2.8 2.8 2.4 16 12 9 4.4 6.2 5 ...
#>  $ TTSW  : num  0.5 0.7 1.4 2.4 4 ...
#>  $ TTPW  : num  2.4 5.4 4 4 21 19 13.2 7.2 9 9 ...
#>  $ AVPW  : num  1.2 1.35 2 4 5.25 4.75 3.3 2.4 1.8 2.25 ...
#>  $ ARSR  : num  2 0 2 0 3 0 0 6 0 0 ...
#>  $ SRDM  : num  42 39.8 29.7 43 37.9 37 38.9 36.9 41 37.9 ...

diversity.compare(x = pdata$CUAL, group = pdata$LNGS, R = 100,
                  base = exp(1), na.omit = TRUE)
#> Computing diversity indices.
#> Performing global permutation tests.
#> 
#> Performing pairwise permutation tests.
#> Computing bootstrap confidence intervals.
#> Generating diversity profiles.
#> $`Diversity Indices`
#> # A tibble: 4 × 30
#>   group   richness margalef_index menhinick_index berger_parker
#>   <chr>      <int>          <dbl>           <dbl>         <dbl>
#> 1 Overall        4          0.585           0.309         0.530
#> 2 Long           3          0.468           0.354         0.514
#> 3 Medium         4          0.701           0.471         0.569
#> 4 Short          4          0.944           0.816         0.458
#> # ℹ 25 more variables: berger_parker_reciprocal <dbl>, simpson <dbl>,
#> #   gini_simpson <dbl>, simpson_max <dbl>, simpson_reciprocal <dbl>,
#> #   simpson_relative <dbl>, simpson_evenness <dbl>, shannon <dbl>,
#> #   shannon_max <dbl>, shannon_relative <dbl>, shannon_ens <dbl>,
#> #   heip_evenness <dbl>, mcintosh_diversity <dbl>, mcintosh_evenness <dbl>,
#> #   smith_wilson <dbl>, brillouin_index <dbl>, renyi_entropy_0 <dbl>,
#> #   renyi_entropy_1 <dbl>, renyi_entropy_2 <dbl>, tsallis_entropy_0 <dbl>, …
#> 
#> $`Global Test`
#>          Measure margalef_index menhinick_index berger_parker
#> 1 Test statistic     4.62557535      3.85787925     0.2539683
#> 2        p-value     0.08910891      0.06930693     0.6534653
#>   berger_parker_reciprocal   simpson gini_simpson simpson_max simpson_relative
#> 1                3.5485816 0.1459742    0.1459742   0.2857143        0.6397893
#> 2                0.5643564 0.5247525    0.4851485   0.0990099        0.2277228
#>     shannon shannon_max shannon_relative shannon_ens heip_evenness
#> 1 2.1255498   7.0871092       1.15089028  55.1369748    10.2305896
#> 2 0.3069307   0.0990099       0.03960396   0.2673267     0.0990099
#>   mcintosh_diversity mcintosh_evenness smith_wilson brillouin_index
#> 1          0.3514116         1.0809743   3.78072900       0.2103725
#> 2          0.2178218         0.1485149   0.00990099       0.8316832
#> 
#> $`Pairwise Test`
#> $`Pairwise Test`$`p-value`
#>        Comparison margalef_index menhinick_index berger_parker
#> 1  Long vs Medium     1.00000000      1.00000000             1
#> 2   Long vs Short     0.02970297      0.02970297             1
#> 3 Medium vs Short     1.00000000      1.00000000             1
#>   berger_parker_reciprocal   simpson gini_simpson simpson_max simpson_relative
#> 1                        1 1.0000000     1.000000   1.0000000        1.0000000
#> 2                        1 0.8613861     1.000000   0.8316832        1.0000000
#> 3                        1 1.0000000     0.950495   1.0000000        0.8910891
#>     shannon shannon_max shannon_relative shannon_ens heip_evenness
#> 1 1.0000000   1.0000000        0.9801980   1.0000000     1.0000000
#> 2 0.2970297   0.7128713        1.0000000   0.3267327     1.0000000
#> 3 0.6831683   1.0000000        0.4752475   0.5346535     0.5346535
#>   mcintosh_diversity mcintosh_evenness smith_wilson brillouin_index
#> 1          1.0000000         0.7425743    1.0000000               1
#> 2          0.4455446         1.0000000    1.0000000               1
#> 3          0.3564356         0.6237624    0.1188119               1
#> 
#> $`Pairwise Test`$cld
#>    Group margalef_index menhinick_index berger_parker berger_parker_reciprocal
#> 1   Long              a               a             a                        a
#> 2 Medium             ab              ab             a                        a
#> 3  Short              b               b             a                        a
#>   simpson gini_simpson simpson_max simpson_relative shannon shannon_max
#> 1       a            a           a                a       a           a
#> 2       a            a           a                a       a           a
#> 3       a            a           a                a       a           a
#>   shannon_relative shannon_ens heip_evenness mcintosh_diversity
#> 1                a           a             a                  a
#> 2                a           a             a                  a
#> 3                a           a             a                  a
#>   mcintosh_evenness smith_wilson brillouin_index
#> 1                 a            a               a
#> 2                 a            a               a
#> 3                 a            a               a
#> 
#> 
#> $`Bootstrap CIs`
#>        Group-CI margalef_index menhinick_index berger_parker
#> 1   Long: lower      0.4676540       0.3535534     0.4027778
#> 2   Long: upper      0.4676540       0.3535534     0.6388889
#> 3 Medium: lower      0.4676540       0.3535534     0.4451389
#> 4 Medium: upper      0.7014810       0.4714045     0.7010417
#> 5  Short: lower      0.6293160       0.6123724     0.3333333
#> 6  Short: upper      0.9439739       0.8164966     0.6666667
#>   berger_parker_reciprocal   simpson gini_simpson simpson_max simpson_relative
#> 1                 1.619091 0.3573495    0.5207948   0.6666667        0.7727431
#> 2                 2.400000 0.4783372    0.6525752   0.6666667        0.9800637
#> 3                 1.471500 0.3587674    0.4611304   0.6666667        0.6932999
#> 4                 2.284476 0.5274981    0.6519965   0.7500000        0.9377894
#> 5                 1.500000 0.2673611    0.4514757   0.6666667        0.6968027
#> 6                 3.000000 0.4874132    0.7395833   0.7500000        0.9861111
#>    shannon shannon_max shannon_relative shannon_ens heip_evenness
#> 1 1.268182    1.584963        0.8026677    3.651240     1.2834996
#> 2 1.550426    1.584963        0.9822709    4.723640     1.8458488
#> 3 1.211598    1.584963        0.6511038    3.348900     0.9269083
#> 4 1.681684    2.000000        0.9404166    5.256166     1.7534524
#> 5 1.199287    1.584963        0.6862734    3.586009     0.9742734
#> 6 1.928964    2.000000        0.9862707    7.085143     2.0408130
#>   mcintosh_diversity mcintosh_evenness smith_wilson brillouin_index
#> 1          0.3468894         0.7185593    0.5133413       0.8075800
#> 2          0.4663058         0.9768175    0.8094535       1.0112951
#> 3          0.3182851         0.6125709    0.3527128       0.7626342
#> 4          0.4622680         0.9136435    0.7572221       1.0709931
#> 5          0.3521915         0.6131844    0.4439441       0.7965172
#> 6          0.6067909         0.9754167    0.8313911       1.1432544
#> 
#> $`Diversity profiles`
#> $`Diversity profiles`$hill
#> $`Diversity profiles`$hill$Long
#>      q observed     mean    lower    upper
#> 1  0.0 3.000000 3.000000 3.000000 3.000000
#> 2  0.1 2.970041 2.964507 2.913407 2.992622
#> 3  0.2 2.940750 2.930167 2.830548 2.985350
#> 4  0.3 2.912153 2.896998 2.751983 2.978185
#> 5  0.4 2.884272 2.865008 2.680495 2.971129
#> 6  0.5 2.857124 2.834195 2.614262 2.964182
#> 7  0.6 2.830721 2.804551 2.553018 2.957346
#> 8  0.7 2.805073 2.776061 2.496473 2.950622
#> 9  0.8 2.780186 2.748702 2.444321 2.944010
#> 10 0.9 2.756060 2.722450 2.396682 2.937511
#> 11 1.0 2.732695 2.697274 2.352870 2.931125
#> 12 1.1 2.710085 2.673144 2.312557 2.924853
#> 13 1.2 2.688224 2.650026 2.275449 2.918694
#> 14 1.3 2.667101 2.627884 2.241266 2.912648
#> 15 1.4 2.646704 2.606682 2.209749 2.906716
#> 16 1.5 2.627019 2.586383 2.180659 2.900895
#> 17 1.6 2.608031 2.566953 2.153775 2.895187
#> 18 1.7 2.589722 2.548353 2.128895 2.889590
#> 19 1.8 2.572075 2.530549 2.105837 2.884104
#> 20 1.9 2.555070 2.513506 2.079712 2.878727
#> 21 2.0 2.538688 2.497189 2.054650 2.873458
#> 22 2.1 2.522908 2.481565 2.031210 2.868297
#> 23 2.2 2.507711 2.466602 2.009283 2.863243
#> 24 2.3 2.493075 2.452270 1.988766 2.858293
#> 25 2.4 2.478981 2.438538 1.969563 2.853447
#> 26 2.5 2.465409 2.425378 1.951581 2.848703
#> 27 2.6 2.452337 2.412763 1.934733 2.844060
#> 28 2.7 2.439748 2.400666 1.918939 2.839516
#> 29 2.8 2.427621 2.389062 1.904123 2.835070
#> 30 2.9 2.415938 2.377927 1.890216 2.830720
#> 31 3.0 2.404680 2.367239 1.877151 2.826465
#> 
#> $`Diversity profiles`$hill$Medium
#>      q observed     mean    lower    upper
#> 1  0.0 4.000000 3.720000 3.000000 4.000000
#> 2  0.1 3.775167 3.554848 2.917060 3.878659
#> 3  0.2 3.586424 3.411515 2.838319 3.765974
#> 4  0.3 3.427791 3.286906 2.763841 3.661485
#> 5  0.4 3.293906 3.178177 2.693637 3.564688
#> 6  0.5 3.180150 3.082815 2.627668 3.475057
#> 7  0.6 3.082659 2.998661 2.565856 3.392056
#> 8  0.7 2.998274 2.923904 2.508087 3.315162
#> 9  0.8 2.924454 2.857039 2.454214 3.248289
#> 10 0.9 2.859178 2.796834 2.400441 3.186299
#> 11 1.0 2.800853 2.742283 2.348804 3.126733
#> 12 1.1 2.748230 2.692567 2.300256 3.063674
#> 13 1.2 2.700330 2.647019 2.252135 3.014408
#> 14 1.3 2.656390 2.605094 2.200321 2.976618
#> 15 1.4 2.615811 2.566344 2.152981 2.943430
#> 16 1.5 2.578125 2.530400 2.109680 2.922474
#> 17 1.6 2.542961 2.496955 2.070033 2.903089
#> 18 1.7 2.510025 2.465752 2.033698 2.884995
#> 19 1.8 2.479080 2.436571 2.000366 2.867969
#> 20 1.9 2.449934 2.409227 1.969761 2.851665
#> 21 2.0 2.422430 2.383558 1.941635 2.833472
#> 22 2.1 2.396434 2.359423 1.915760 2.816125
#> 23 2.2 2.371835 2.336700 1.891934 2.799526
#> 24 2.3 2.348536 2.315280 1.869970 2.783596
#> 25 2.4 2.326451 2.295065 1.849702 2.768270
#> 26 2.5 2.305505 2.275967 1.830976 2.753492
#> 27 2.6 2.285629 2.257907 1.813656 2.739219
#> 28 2.7 2.266761 2.240814 1.797617 2.725413
#> 29 2.8 2.248843 2.224623 1.782746 2.712043
#> 30 2.9 2.231821 2.209274 1.768940 2.699083
#> 31 3.0 2.215647 2.194711 1.756108 2.686511
#> 
#> $`Diversity profiles`$hill$Short
#>      q observed     mean    lower    upper
#> 1  0.0 4.000000 3.920000 3.000000 4.000000
#> 2  0.1 3.933810 3.834336 2.921506 3.988856
#> 3  0.2 3.870196 3.753673 2.846484 3.977595
#> 4  0.3 3.809136 3.677884 2.775035 3.966257
#> 5  0.4 3.750595 3.606797 2.707211 3.955115
#> 6  0.5 3.694521 3.540210 2.643022 3.944035
#> 7  0.6 3.640856 3.477895 2.582436 3.933020
#> 8  0.7 3.589527 3.419614 2.525386 3.922074
#> 9  0.8 3.540459 3.365120 2.471777 3.911201
#> 10 0.9 3.493567 3.314170 2.421490 3.900405
#> 11 1.0 3.448767 3.266524 2.351508 3.889690
#> 12 1.1 3.405970 3.221950 2.282823 3.879059
#> 13 1.2 3.365087 3.180229 2.220895 3.868516
#> 14 1.3 3.326031 3.141152 2.165081 3.858063
#> 15 1.4 3.288713 3.104524 2.114774 3.847703
#> 16 1.5 3.253050 3.070164 2.070403 3.837438
#> 17 1.6 3.218958 3.037901 2.030373 3.827273
#> 18 1.7 3.186359 3.007579 1.994125 3.817208
#> 19 1.8 3.155176 2.979054 1.961252 3.807245
#> 20 1.9 3.125338 2.952193 1.931386 3.797388
#> 21 2.0 3.096774 2.926872 1.904202 3.787636
#> 22 2.1 3.069420 2.902980 1.879410 3.777993
#> 23 2.2 3.043214 2.880412 1.856755 3.768459
#> 24 2.3 3.018098 2.859073 1.836008 3.759036
#> 25 2.4 2.994016 2.838875 1.816971 3.749724
#> 26 2.5 2.970917 2.819739 1.799467 3.740525
#> 27 2.6 2.948751 2.801590 1.783340 3.731439
#> 28 2.7 2.927474 2.784361 1.768452 3.722468
#> 29 2.8 2.907041 2.767989 1.754681 3.713611
#> 30 2.9 2.887412 2.752418 1.741920 3.704868
#> 31 3.0 2.868549 2.737593 1.730072 3.696241
#> 
#> attr(,"R")
#> [1] 100
#> attr(,"conf")
#> [1] 0.95
#> attr(,"parameter")
#> [1] "hill"
#> attr(,"ci.type")
#> [1] "perc"
#> 
#> $`Diversity profiles`$renyi
#> $`Diversity profiles`$renyi$Long
#>      q  observed      mean     lower    upper
#> 1  0.0 1.0986123 1.0986123 1.0986123 1.098612
#> 2  0.1 1.0885756 1.0867065 1.0716769 1.096389
#> 3  0.2 1.0786646 1.0749957 1.0450036 1.094156
#> 4  0.3 1.0688928 1.0635027 1.0187053 1.091913
#> 5  0.4 1.0592726 1.0522473 0.9928892 1.089662
#> 6  0.5 1.0498155 1.0412470 0.9676551 1.087404
#> 7  0.6 1.0405315 1.0305164 0.9430928 1.085140
#> 8  0.7 1.0314297 1.0200672 0.9192806 1.082870
#> 9  0.8 1.0225177 1.0099086 0.8962842 1.080596
#> 10 0.9 1.0138021 1.0000474 0.8741559 1.078319
#> 11 1.0 1.0052882 0.9904876 0.8529346 1.076038
#> 12 1.1 0.9969801 0.9812315 0.8326461 1.073757
#> 13 1.2 0.9888807 0.9722792 0.8135935 1.071474
#> 14 1.3 0.9809920 0.9636289 0.7960083 1.069192
#> 15 1.4 0.9733151 0.9552777 0.7793783 1.066912
#> 16 1.5 0.9658499 0.9472210 0.7636711 1.064634
#> 17 1.6 0.9585956 0.9394532 0.7488499 1.062359
#> 18 1.7 0.9515507 0.9319679 0.7348751 1.060089
#> 19 1.8 0.9447130 0.9247578 0.7217054 1.057824
#> 20 1.9 0.9380796 0.9178153 0.7092988 1.055565
#> 21 2.0 0.9316472 0.9111322 0.6976128 1.053313
#> 22 2.1 0.9254122 0.9046999 0.6855324 1.051311
#> 23 2.2 0.9193702 0.8985098 0.6741376 1.049335
#> 24 2.3 0.9135169 0.8925532 0.6633944 1.047385
#> 25 2.4 0.9078477 0.8868214 0.6532689 1.045462
#> 26 2.5 0.9023576 0.8813057 0.6437270 1.043564
#> 27 2.6 0.8970416 0.8759974 0.6347355 1.041693
#> 28 2.7 0.8918947 0.8708883 0.6262620 1.039846
#> 29 2.8 0.8869118 0.8659701 0.6182754 1.038026
#> 30 2.9 0.8820876 0.8612349 0.6107455 1.036231
#> 31 3.0 0.8774170 0.8566748 0.6036439 1.034460
#> 
#> $`Diversity profiles`$renyi$Medium
#>      q  observed      mean     lower    upper
#> 1  0.0 1.3862944 1.2712215 1.0986123 1.386294
#> 2  0.1 1.3284446 1.2354654 1.0754906 1.356920
#> 3  0.2 1.2771555 1.2023144 1.0524286 1.328999
#> 4  0.3 1.2319160 1.1716614 1.0295155 1.302535
#> 5  0.4 1.1920742 1.1433397 1.0068396 1.277477
#> 6  0.5 1.1569284 1.1171510 0.9844864 1.253727
#> 7  0.6 1.1257925 1.0928886 0.9625369 1.231286
#> 8  0.7 1.0980367 1.0703514 0.9410657 1.210094
#> 9  0.8 1.0731077 1.0493541 0.9201397 1.190085
#> 10 0.9 1.0505342 1.0297308 0.8998173 1.171190
#> 11 1.0 1.0299241 1.0113363 0.8801472 1.153339
#> 12 1.1 1.0109570 0.9940456 0.8611687 1.136461
#> 13 1.2 0.9933740 0.9777519 0.8398910 1.121734
#> 14 1.3 0.9769680 0.9623642 0.8183980 1.109621
#> 15 1.4 0.9615743 0.9478051 0.7982844 1.098325
#> 16 1.5 0.9470625 0.9340086 0.7796244 1.090874
#> 17 1.6 0.9333293 0.9209177 0.7621082 1.084958
#> 18 1.7 0.9202928 0.9084831 0.7456683 1.077946
#> 19 1.8 0.9078876 0.8966617 0.7302405 1.067209
#> 20 1.9 0.8960613 0.8854153 0.7157642 1.057666
#> 21 2.0 0.8847711 0.8747096 0.7021816 1.051146
#> 22 2.1 0.8739819 0.8645137 0.6894378 1.044707
#> 23 2.2 0.8636641 0.8547994 0.6774802 1.038548
#> 24 2.3 0.8537922 0.8455406 0.6662590 1.032869
#> 25 2.4 0.8443440 0.8367132 0.6557266 1.028852
#> 26 2.5 0.8352998 0.8282945 0.6458380 1.024978
#> 27 2.6 0.8266413 0.8202633 0.6365508 1.021234
#> 28 2.7 0.8183519 0.8125996 0.6278246 1.017608
#> 29 2.8 0.8104158 0.8052847 0.6196218 1.014093
#> 30 2.9 0.8028180 0.7983005 0.6119070 1.010680
#> 31 3.0 0.7955444 0.7916303 0.6046468 1.007362
#> 
#> $`Diversity profiles`$renyi$Short
#>      q observed      mean     lower    upper
#> 1  0.0 1.386294 1.3517725 1.0986123 1.386294
#> 2  0.1 1.369609 1.3302319 1.0745579 1.380347
#> 3  0.2 1.353305 1.3093484 1.0503188 1.374583
#> 4  0.3 1.337402 1.2891692 1.0260126 1.369001
#> 5  0.4 1.321914 1.2697285 1.0017614 1.363601
#> 6  0.5 1.306851 1.2510482 0.9776878 1.358383
#> 7  0.6 1.292219 1.2331396 0.9539118 1.353343
#> 8  0.7 1.278021 1.2160045 0.9305471 1.348480
#> 9  0.8 1.264256 1.1996371 0.9076982 1.343791
#> 10 0.9 1.250923 1.1840250 0.8854578 1.339272
#> 11 1.0 1.238017 1.1691507 0.8639052 1.334920
#> 12 1.1 1.225530 1.1549928 0.8431051 1.330730
#> 13 1.2 1.213454 1.1415269 0.8165183 1.326652
#> 14 1.3 1.201780 1.1287264 0.7909496 1.322228
#> 15 1.4 1.190496 1.1165637 0.7671094 1.317886
#> 16 1.5 1.179593 1.1050098 0.7449247 1.313625
#> 17 1.6 1.169058 1.0940359 0.7243118 1.309446
#> 18 1.7 1.158879 1.0836130 0.7051809 1.305345
#> 19 1.8 1.149044 1.0737128 0.6874391 1.301324
#> 20 1.9 1.139542 1.0643073 0.6709929 1.297380
#> 21 2.0 1.130361 1.0553696 0.6557503 1.293513
#> 22 2.1 1.121489 1.0468738 0.6416223 1.290362
#> 23 2.2 1.112914 1.0387949 0.6285237 1.287322
#> 24 2.3 1.104627 1.0311092 0.6163742 1.284387
#> 25 2.4 1.096616 1.0237942 0.6050982 1.281554
#> 26 2.5 1.088871 1.0168283 0.5946253 1.278818
#> 27 2.6 1.081382 1.0101915 0.5848902 1.276176
#> 28 2.7 1.074140 1.0038644 0.5758327 1.273622
#> 29 2.8 1.067136 0.9978293 0.5673973 1.271154
#> 30 2.9 1.060361 0.9920691 0.5595332 1.268767
#> 31 3.0 1.053806 0.9865680 0.5521937 1.266458
#> 
#> attr(,"R")
#> [1] 100
#> attr(,"conf")
#> [1] 0.95
#> attr(,"parameter")
#> [1] "renyi"
#> attr(,"ci.type")
#> [1] "perc"
#> 
#> $`Diversity profiles`$tsallis
#> $`Diversity profiles`$tsallis$Long
#>      q  observed      mean     lower     upper
#> 1  0.0 2.0000000 2.0000000 2.0000000 2.0000000
#> 2  0.1 1.8485612 1.8430441 1.7910575 1.8700431
#> 3  0.2 1.7126236 1.7031448 1.6149316 1.7506878
#> 4  0.3 1.5903508 1.5781030 1.4655903 1.6409685
#> 5  0.4 1.4801432 1.4660366 1.3351151 1.5400141
#> 6  0.5 1.3806058 1.3653308 1.2219832 1.4470380
#> 7  0.6 1.2905205 1.2745974 1.1235738 1.3613299
#> 8  0.7 1.2088223 1.1926393 1.0375969 1.2822476
#> 9  0.8 1.1345790 1.1184225 0.9621552 1.2092105
#> 10 0.9 1.0669734 1.0510515 0.8958623 1.1416928
#> 11 1.0 1.0052882 0.9897487 0.8372554 1.0792184
#> 12 1.1 0.9488928 0.9338379 0.7851750 1.0213560
#> 13 1.2 0.8972324 0.8827291 0.7386987 0.9677144
#> 14 1.3 0.8498176 0.8359070 0.6970527 0.9179385
#> 15 1.4 0.8062167 0.7929199 0.6595853 0.8717060
#> 16 1.5 0.7660477 0.7533715 0.6257467 0.8287239
#> 17 1.6 0.7289728 0.7169132 0.5950708 0.7887259
#> 18 1.7 0.6946920 0.6832376 0.5671616 0.7514697
#> 19 1.8 0.6629392 0.6520730 0.5416815 0.7167346
#> 20 1.9 0.6334774 0.6231791 0.5183417 0.6843196
#> 21 2.0 0.6060957 0.5963426 0.4968943 0.6540413
#> 22 2.1 0.5806057 0.5713739 0.4771259 0.6257325
#> 23 2.2 0.5568392 0.5481041 0.4588522 0.5992404
#> 24 2.3 0.5346455 0.5263826 0.4419134 0.5744256
#> 25 2.4 0.5138895 0.5060746 0.4262648 0.5511604
#> 26 2.5 0.4944499 0.4870595 0.4120504 0.5293282
#> 27 2.6 0.4762176 0.4692288 0.3987271 0.5088220
#> 28 2.7 0.4590941 0.4524852 0.3862123 0.4895440
#> 29 2.8 0.4429909 0.4367409 0.3744335 0.4714044
#> 30 2.9 0.4278278 0.4219168 0.3633266 0.4543207
#> 31 3.0 0.4135320 0.4079413 0.3528350 0.4382174
#> 
#> $`Diversity profiles`$tsallis$Medium
#>      q  observed      mean     lower     upper
#> 1  0.0 3.0000000 2.6200000 2.0000000 3.0000000
#> 2  0.1 2.5617123 2.3001691 1.8070878 2.6641696
#> 3  0.2 2.2224763 2.0400464 1.6383924 2.3820368
#> 4  0.3 1.9552854 1.8257399 1.4905924 2.1432938
#> 5  0.4 1.7411443 1.6470172 1.3608427 1.9398361
#> 6  0.5 1.5665950 1.4962700 1.2467044 1.7652446
#> 7  0.6 1.4220235 1.3677850 1.1460846 1.6144161
#> 8  0.7 1.3004971 1.2572269 1.0571861 1.4810648
#> 9  0.8 1.1969637 1.1612703 0.9784642 1.3638935
#> 10 0.9 1.1076994 1.0773365 0.9085899 1.2608095
#> 11 1.0 1.0299241 1.0034035 0.8464190 1.1697039
#> 12 1.1 0.9615347 0.9378685 0.7883323 1.0888291
#> 13 1.2 0.9009177 0.8794473 0.7342359 1.0167313
#> 14 1.3 0.8468176 0.8271002 0.6866586 0.9527941
#> 15 1.4 0.7982434 0.7799771 0.6446108 0.8954410
#> 16 1.5 0.7544018 0.7373768 0.6072742 0.8436423
#> 17 1.6 0.7146493 0.6987154 0.5739693 0.7966894
#> 18 1.7 0.6784574 0.6635031 0.5441299 0.7539833
#> 19 1.8 0.6453869 0.6313260 0.5172818 0.7151080
#> 20 1.9 0.6150691 0.6018316 0.4930261 0.6798456
#> 21 2.0 0.5871914 0.5747184 0.4710262 0.6474248
#> 22 2.1 0.5614865 0.5497265 0.4509970 0.6175388
#> 23 2.2 0.5377246 0.5266314 0.4326958 0.5904820
#> 24 2.3 0.5157061 0.5052374 0.4159152 0.5653574
#> 25 2.4 0.4952573 0.4853739 0.4004778 0.5419253
#> 26 2.5 0.4762260 0.4668914 0.3862306 0.5200429
#> 27 2.6 0.4584783 0.4496582 0.3730419 0.4996052
#> 28 2.7 0.4418959 0.4335582 0.3607975 0.4804871
#> 29 2.8 0.4263738 0.4184885 0.3493982 0.4625766
#> 30 2.9 0.4118190 0.4043577 0.3387576 0.4457736
#> 31 3.0 0.3981481 0.3910846 0.3288002 0.4299877
#> 
#> $`Diversity profiles`$tsallis$Short
#>      q  observed      mean     lower     upper
#> 1  0.0 3.0000000 2.8800000 2.0000000 3.0000000
#> 2  0.1 2.7003331 2.5734826 1.8147007 2.7421172
#> 3  0.2 2.4405950 2.3124385 1.6541673 2.5116032
#> 4  0.3 2.2146076 2.0888371 1.5145195 2.3052466
#> 5  0.4 2.0172417 1.8962276 1.3925377 2.1202369
#> 6  0.5 1.8442276 1.7294050 1.2855441 1.9541124
#> 7  0.6 1.6920012 1.5841507 1.1913073 1.8047151
#> 8  0.7 1.5575795 1.4570295 1.1081597 1.6701513
#> 9  0.8 1.4384586 1.3452315 1.0343970 1.5487575
#> 10 0.9 1.3325309 1.2464472 0.9698694 1.4390715
#> 11 1.0 1.2380168 1.1587697 0.9129924 1.3398061
#> 12 1.1 1.1534096 1.0806170 0.8619297 1.2498275
#> 13 1.2 1.0774297 1.0106708 0.8158643 1.1681359
#> 14 1.3 1.0089869 0.9478270 0.7741143 1.0938485
#> 15 1.4 0.9471496 0.8911572 0.7352384 1.0261854
#> 16 1.5 0.8911198 0.8398767 0.6960151 0.9644565
#> 17 1.6 0.8402110 0.7933200 0.6608967 0.9080507
#> 18 1.7 0.7938317 0.7509197 0.6290807 0.8564263
#> 19 1.8 0.7514702 0.7121901 0.5992307 0.8091025
#> 20 1.9 0.7126828 0.6767142 0.5701165 0.7656519
#> 21 2.0 0.6770833 0.6441319 0.5436632 0.7256944
#> 22 2.1 0.6443352 0.6141317 0.5195342 0.6888917
#> 23 2.2 0.6141439 0.5864426 0.4974444 0.6549420
#> 24 2.3 0.5862509 0.5608282 0.4771513 0.6235757
#> 25 2.4 0.5604291 0.5370818 0.4584475 0.5945523
#> 26 2.5 0.5364780 0.5150217 0.4411551 0.5676562
#> 27 2.6 0.5142206 0.4944880 0.4251208 0.5426946
#> 28 2.7 0.4934997 0.4753392 0.4102120 0.5194943
#> 29 2.8 0.4741760 0.4574500 0.3963135 0.4979000
#> 30 2.9 0.4561251 0.4407092 0.3833251 0.4777719
#> 31 3.0 0.4392361 0.4250174 0.3711589 0.4589844
#> 
#> attr(,"R")
#> [1] 100
#> attr(,"conf")
#> [1] 0.95
#> attr(,"parameter")
#> [1] "tsallis"
#> attr(,"ci.type")
#> [1] "perc"
#> 
#> 

diversity.compare(x = pdata$ANGB, group = pdata$LNGS, R = 100,
                  base = exp(1), na.omit = TRUE)
#> Computing diversity indices.
#> Performing global permutation tests.
#> 
#> Performing pairwise permutation tests.
#> Computing bootstrap confidence intervals.
#> Generating diversity profiles.
#> $`Diversity Indices`
#> # A tibble: 4 × 30
#>   group   richness margalef_index menhinick_index berger_parker
#>   <chr>      <int>          <dbl>           <dbl>         <dbl>
#> 1 Overall        4          0.585           0.309         0.452
#> 2 Long           4          0.701           0.471         0.514
#> 3 Medium         4          0.701           0.471         0.444
#> 4 Short          3          0.629           0.612         0.458
#> # ℹ 25 more variables: berger_parker_reciprocal <dbl>, simpson <dbl>,
#> #   gini_simpson <dbl>, simpson_max <dbl>, simpson_reciprocal <dbl>,
#> #   simpson_relative <dbl>, simpson_evenness <dbl>, shannon <dbl>,
#> #   shannon_max <dbl>, shannon_relative <dbl>, shannon_ens <dbl>,
#> #   heip_evenness <dbl>, mcintosh_diversity <dbl>, mcintosh_evenness <dbl>,
#> #   smith_wilson <dbl>, brillouin_index <dbl>, renyi_entropy_0 <dbl>,
#> #   renyi_entropy_1 <dbl>, renyi_entropy_2 <dbl>, tsallis_entropy_0 <dbl>, …
#> 
#> $`Global Test`
#>          Measure margalef_index menhinick_index berger_parker
#> 1 Test statistic      0.1071317       0.4087945     0.1825397
#> 2        p-value      1.0000000       1.0000000     0.5643564
#>   berger_parker_reciprocal    simpson gini_simpson simpson_max simpson_relative
#> 1                3.4727768 0.03859249   0.03859249   0.1428571        0.2681430
#> 2                0.6633663 0.76237624   0.65346535   0.2574257        0.2574257
#>     shannon shannon_max shannon_relative shannon_ens heip_evenness
#> 1 0.6662977   3.5435546        0.3615327  17.6011822     3.0824569
#> 2 0.7029703   0.3168317        0.3069307   0.6534653     0.3663366
#>   mcintosh_diversity mcintosh_evenness smith_wilson brillouin_index
#> 1         0.07321862         0.4906392    1.6496292       0.6374954
#> 2         0.63366337         0.2277228    0.1386139       0.4851485
#> 
#> $`Pairwise Test`
#> $`Pairwise Test`$`p-value`
#>        Comparison margalef_index menhinick_index berger_parker
#> 1  Long vs Medium              1               1             1
#> 2   Long vs Short              1               1             1
#> 3 Medium vs Short              1               1             1
#>   berger_parker_reciprocal simpson gini_simpson simpson_max simpson_relative
#> 1                        1       1            1   1.0000000        1.0000000
#> 2                        1       1            1   1.0000000        0.5346535
#> 3                        1       1            1   0.8613861        0.5940594
#>   shannon shannon_max shannon_relative shannon_ens heip_evenness
#> 1       1   1.0000000        1.0000000           1     1.0000000
#> 2       1   0.9207921        0.2079208           1     0.5643564
#> 3       1   0.9801980        0.6534653           1     0.8019802
#>   mcintosh_diversity mcintosh_evenness smith_wilson brillouin_index
#> 1                  1         1.0000000    1.0000000       1.0000000
#> 2                  1         0.4158416    0.1485149       1.0000000
#> 3                  1         0.5643564    0.3861386       0.9207921
#> 
#> $`Pairwise Test`$cld
#>    Group margalef_index menhinick_index berger_parker berger_parker_reciprocal
#> 1   Long              a               a             a                        a
#> 2 Medium              a               a             a                        a
#> 3  Short              a               a             a                        a
#>   simpson gini_simpson simpson_max simpson_relative shannon shannon_max
#> 1       a            a           a                a       a           a
#> 2       a            a           a                a       a           a
#> 3       a            a           a                a       a           a
#>   shannon_relative shannon_ens heip_evenness mcintosh_diversity
#> 1                a           a             a                  a
#> 2                a           a             a                  a
#> 3                a           a             a                  a
#>   mcintosh_evenness smith_wilson brillouin_index
#> 1                 a            a               a
#> 2                 a            a               a
#> 3                 a            a               a
#> 
#> 
#> $`Bootstrap CIs`
#>        Group-CI margalef_index menhinick_index berger_parker
#> 1   Long: lower       0.467654       0.4714045     0.4034722
#> 2   Long: upper       0.701481       0.4714045     0.6170139
#> 3 Medium: lower       0.467654       0.4154252     0.3611111
#> 4 Medium: upper       0.701481       0.4714045     0.5687500
#> 5  Short: lower       0.629316       0.6123724     0.3750000
#> 6  Short: upper       0.629316       0.6123724     0.6447917
#>   berger_parker_reciprocal   simpson gini_simpson simpson_max simpson_relative
#> 1                 1.516755 0.3092882    0.5582658   0.7500000        0.7493699
#> 2                 2.482759 0.4516590    0.6944252   0.7500000        0.9397377
#> 3                 1.783516 0.2961323    0.5848476   0.7500000        0.7832176
#> 4                 2.821846 0.4153935    0.7056327   0.7500000        0.9402392
#> 5                 1.411765 0.3368056    0.4548611   0.6666667        0.7069010
#> 6                 2.666667 0.4618056    0.6648437   0.6666667        0.9972656
#>    shannon shannon_max shannon_relative shannon_ens heip_evenness
#> 1 1.398932    2.000000        0.6846781    3.849376     0.9445944
#> 2 1.829966    2.000000        0.9186726    6.089423     1.6918448
#> 3 1.463025    2.000000        0.7325934    4.475436     1.1927354
#> 4 1.870477    2.000000        0.9164158    6.326070     1.8351325
#> 5 1.207261    1.584963        0.7816209    3.320279     1.1066169
#> 6 1.584963    1.584963        0.9975045    4.879108     1.9212443
#>   mcintosh_diversity mcintosh_evenness smith_wilson brillouin_index
#> 1          0.3676601         0.6350560    0.3494951       0.8977944
#> 2          0.5069574         0.8860126    0.6109161       1.1644251
#> 3          0.4156116         0.7131378    0.3557487       0.9646353
#> 4          0.5215998         0.9381250    0.6856812       1.1923234
#> 5          0.3477589         0.6548518    0.4945224       0.6049201
#> 6          0.5290714         0.9929037    0.9203697       0.9522141
#> 
#> $`Diversity profiles`
#> $`Diversity profiles`$hill
#> $`Diversity profiles`$hill$Long
#>      q observed     mean    lower    upper
#> 1  0.0 4.000000 3.970000 3.000000 4.000000
#> 2  0.1 3.886394 3.844652 2.957442 3.937434
#> 3  0.2 3.779642 3.729528 2.915691 3.876881
#> 4  0.3 3.679590 3.623901 2.874819 3.818355
#> 5  0.4 3.586026 3.527025 2.834893 3.761858
#> 6  0.5 3.498684 3.438169 2.795973 3.708884
#> 7  0.6 3.417265 3.356631 2.758112 3.660657
#> 8  0.7 3.341444 3.281752 2.721355 3.615361
#> 9  0.8 3.270881 3.212920 2.685739 3.572695
#> 10 0.9 3.205233 3.149575 2.651292 3.532420
#> 11 1.0 3.144158 3.091204 2.618037 3.491691
#> 12 1.1 3.087324 3.037343 2.585987 3.451114
#> 13 1.2 3.034412 2.987573 2.535192 3.412174
#> 14 1.3 2.985118 2.941514 2.487798 3.374805
#> 15 1.4 2.939157 2.898824 2.444774 3.338766
#> 16 1.5 2.896264 2.859196 2.396583 3.307408
#> 17 1.6 2.856191 2.822354 2.349466 3.281465
#> 18 1.7 2.818711 2.788050 2.309700 3.256929
#> 19 1.8 2.783616 2.756059 2.269781 3.233741
#> 20 1.9 2.750715 2.726183 2.230360 3.211791
#> 21 2.0 2.719832 2.698238 2.196904 3.190979
#> 22 2.1 2.690809 2.672064 2.166077 3.171217
#> 23 2.2 2.663500 2.647513 2.137630 3.152423
#> 24 2.3 2.637774 2.624452 2.111339 3.134524
#> 25 2.4 2.613511 2.602763 2.087007 3.117454
#> 26 2.5 2.590601 2.582338 2.064453 3.101154
#> 27 2.6 2.568944 2.563078 2.043518 3.085569
#> 28 2.7 2.548450 2.544894 2.024058 3.075441
#> 29 2.8 2.529037 2.527708 2.005945 3.066514
#> 30 2.9 2.510629 2.511444 1.989061 3.058039
#> 31 3.0 2.493156 2.496038 1.973303 3.049975
#> 
#> $`Diversity profiles`$hill$Medium
#>      q observed     mean    lower    upper
#> 1  0.0 4.000000 3.990000 4.000000 4.000000
#> 2  0.1 3.902761 3.881627 3.770824 3.961054
#> 3  0.2 3.812649 3.782888 3.577933 3.922901
#> 4  0.3 3.729325 3.692978 3.415458 3.885509
#> 5  0.4 3.652407 3.611096 3.278109 3.848908
#> 6  0.5 3.581486 3.536467 3.161307 3.813117
#> 7  0.6 3.516139 3.468366 3.061207 3.778147
#> 8  0.7 3.455945 3.406127 2.979496 3.744009
#> 9  0.8 3.400487 3.349142 2.927379 3.710709
#> 10 0.9 3.349367 3.296865 2.879631 3.678250
#> 11 1.0 3.302206 3.248808 2.835741 3.646633
#> 12 1.1 3.258648 3.204534 2.796370 3.616290
#> 13 1.2 3.218365 3.163658 2.749966 3.591454
#> 14 1.3 3.181053 3.125836 2.696542 3.567923
#> 15 1.4 3.146434 3.090765 2.645859 3.545617
#> 16 1.5 3.114256 3.058176 2.595767 3.524463
#> 17 1.6 3.084291 3.027829 2.549447 3.504388
#> 18 1.7 3.056333 2.999515 2.506574 3.485325
#> 19 1.8 3.030197 2.973046 2.466851 3.466950
#> 20 1.9 3.005715 2.948254 2.430006 3.446189
#> 21 2.0 2.982739 2.924992 2.395793 3.424042
#> 22 2.1 2.961134 2.903126 2.363989 3.402281
#> 23 2.2 2.940781 2.882540 2.334389 3.381138
#> 24 2.3 2.921573 2.863128 2.306810 3.360600
#> 25 2.4 2.903412 2.844794 2.281142 3.340654
#> 26 2.5 2.886214 2.827453 2.257187 3.321290
#> 27 2.6 2.869900 2.811029 2.234779 3.303679
#> 28 2.7 2.854401 2.795453 2.213795 3.288901
#> 29 2.8 2.839655 2.780662 2.194123 3.274655
#> 30 2.9 2.825604 2.766599 2.175661 3.260914
#> 31 3.0 2.812200 2.753214 2.158318 3.247653
#> 
#> $`Diversity profiles`$hill$Short
#>      q observed     mean    lower    upper
#> 1  0.0 3.000000 3.000000 3.000000 3.000000
#> 2  0.1 2.989707 2.979641 2.941363 2.998426
#> 3  0.2 2.979386 2.959673 2.884567 2.996855
#> 4  0.3 2.969042 2.940115 2.829455 2.995285
#> 5  0.4 2.958682 2.920985 2.775177 2.993718
#> 6  0.5 2.948312 2.902295 2.722420 2.992154
#> 7  0.6 2.937939 2.884057 2.671301 2.990593
#> 8  0.7 2.927570 2.866281 2.621919 2.989034
#> 9  0.8 2.917212 2.848971 2.574355 2.987478
#> 10 0.9 2.906871 2.832133 2.528669 2.985925
#> 11 1.0 2.896553 2.815769 2.484900 2.984375
#> 12 1.1 2.886267 2.799877 2.443071 2.982828
#> 13 1.2 2.876017 2.784456 2.403187 2.981285
#> 14 1.3 2.865812 2.769503 2.365235 2.979745
#> 15 1.4 2.855656 2.755013 2.329338 2.978209
#> 16 1.5 2.845557 2.740978 2.299143 2.976676
#> 17 1.6 2.835521 2.727392 2.270752 2.975148
#> 18 1.7 2.825553 2.714246 2.244062 2.973623
#> 19 1.8 2.815659 2.701531 2.218971 2.972102
#> 20 1.9 2.805845 2.689236 2.192123 2.970585
#> 21 2.0 2.796117 2.677351 2.165414 2.969072
#> 22 2.1 2.786478 2.665865 2.140197 2.967564
#> 23 2.2 2.776935 2.654766 2.116408 2.966060
#> 24 2.3 2.767491 2.644043 2.093980 2.964560
#> 25 2.4 2.758150 2.633684 2.072846 2.963065
#> 26 2.5 2.748918 2.623677 2.052936 2.961575
#> 27 2.6 2.739797 2.614010 2.034185 2.960089
#> 28 2.7 2.730792 2.604671 2.016525 2.958608
#> 29 2.8 2.721904 2.595650 1.999892 2.957132
#> 30 2.9 2.713137 2.586933 1.984225 2.955662
#> 31 3.0 2.704494 2.578511 1.969464 2.954196
#> 
#> attr(,"R")
#> [1] 100
#> attr(,"conf")
#> [1] 0.95
#> attr(,"parameter")
#> [1] "hill"
#> attr(,"ci.type")
#> [1] "perc"
#> 
#> $`Diversity profiles`$renyi
#> $`Diversity profiles`$renyi$Long
#>      q  observed      mean     lower    upper
#> 1  0.0 1.3862944 1.3805407 1.2496454 1.386294
#> 2  0.1 1.3574818 1.3467986 1.2037093 1.371691
#> 3  0.2 1.3296293 1.3146439 1.1613260 1.357354
#> 4  0.3 1.3028014 1.2841134 1.1225260 1.343300
#> 5  0.4 1.2770445 1.2552064 1.0872061 1.329545
#> 6  0.5 1.2523869 1.2278942 1.0551672 1.316100
#> 7  0.6 1.2288406 1.2021287 1.0261530 1.302977
#> 8  0.7 1.2064030 1.1778494 0.9998812 1.290182
#> 9  0.8 1.1850594 1.1549878 0.9760672 1.277720
#> 10 0.9 1.1647848 1.1334715 0.9545664 1.265595
#> 11 1.0 1.1455462 1.1132264 0.9403713 1.253808
#> 12 1.1 1.1273048 1.0941786 0.9226953 1.242359
#> 13 1.2 1.1100177 1.0762557 0.9051133 1.231246
#> 14 1.3 1.0936393 1.0593875 0.8886494 1.220465
#> 15 1.4 1.0781229 1.0435068 0.8732091 1.210012
#> 16 1.5 1.0634215 1.0285493 0.8538808 1.199883
#> 17 1.6 1.0494889 1.0144543 0.8333266 1.190070
#> 18 1.7 1.0362798 1.0011645 0.8148084 1.180567
#> 19 1.8 1.0237510 0.9886261 0.7984615 1.171367
#> 20 1.9 1.0118608 0.9767886 0.7832548 1.162462
#> 21 2.0 1.0005702 0.9656050 0.7690984 1.153844
#> 22 2.1 0.9898419 0.9550314 0.7559092 1.145505
#> 23 2.2 0.9796412 0.9450269 0.7437026 1.137437
#> 24 2.3 0.9699355 0.9355536 0.7324660 1.129631
#> 25 2.4 0.9606945 0.9265763 0.7219631 1.122079
#> 26 2.5 0.9518897 0.9180622 0.7121349 1.114772
#> 27 2.6 0.9434949 0.9099810 0.7029275 1.107077
#> 28 2.7 0.9354855 0.9023045 0.6942916 1.099299
#> 29 2.8 0.9278387 0.8950069 0.6861824 1.091765
#> 30 2.9 0.9205332 0.8880638 0.6785589 1.084469
#> 31 3.0 0.9135494 0.8814529 0.6713337 1.078713
#> 
#> $`Diversity profiles`$renyi$Medium
#>      q observed     mean     lower    upper
#> 1  0.0 1.386294 1.386294 1.3862944 1.386294
#> 2  0.1 1.361684 1.359842 1.3299986 1.375825
#> 3  0.2 1.338324 1.334944 1.2807176 1.365653
#> 4  0.3 1.316227 1.311581 1.2379155 1.355784
#> 5  0.4 1.295386 1.289705 1.2009025 1.346223
#> 6  0.5 1.275778 1.269248 1.1689327 1.336972
#> 7  0.6 1.257364 1.250129 1.1412747 1.328029
#> 8  0.7 1.240096 1.232265 1.1172546 1.319393
#> 9  0.8 1.223919 1.215567 1.0910379 1.311061
#> 10 0.9 1.208771 1.199951 1.0662838 1.303028
#> 11 1.0 1.194591 1.185335 1.0436300 1.295288
#> 12 1.1 1.181312 1.171640 1.0228489 1.287833
#> 13 1.2 1.168873 1.158795 1.0035584 1.280657
#> 14 1.3 1.157212 1.146732 0.9888917 1.273751
#> 15 1.4 1.146270 1.135389 0.9760726 1.267107
#> 16 1.5 1.135990 1.124710 0.9642301 1.260714
#> 17 1.6 1.126322 1.114641 0.9532340 1.254565
#> 18 1.7 1.117216 1.105137 0.9429775 1.248649
#> 19 1.8 1.108628 1.096153 0.9324885 1.242920
#> 20 1.9 1.100516 1.087649 0.9212869 1.237397
#> 21 2.0 1.092842 1.079591 0.9107415 1.232078
#> 22 2.1 1.085572 1.071946 0.9007980 1.226955
#> 23 2.2 1.078675 1.064683 0.8914080 1.222018
#> 24 2.3 1.072122 1.057777 0.8825283 1.217260
#> 25 2.4 1.065887 1.051201 0.8741204 1.212672
#> 26 2.5 1.059945 1.044934 0.8663092 1.208248
#> 27 2.6 1.054277 1.038955 0.8589023 1.204347
#> 28 2.7 1.048862 1.033245 0.8518644 1.201467
#> 29 2.8 1.043682 1.027787 0.8451700 1.198726
#> 30 2.9 1.038722 1.022565 0.8387959 1.196113
#> 31 3.0 1.033967 1.017565 0.8327210 1.193618
#> 
#> $`Diversity profiles`$renyi$Short
#>      q  observed      mean     lower    upper
#> 1  0.0 1.0986123 1.0986123 1.0986123 1.098612
#> 2  0.1 1.0951755 1.0910644 1.0710416 1.098337
#> 3  0.2 1.0917172 1.0835984 1.0437308 1.098062
#> 4  0.3 1.0882393 1.0762277 1.0167984 1.097787
#> 5  0.4 1.0847438 1.0689647 0.9903557 1.097512
#> 6  0.5 1.0812328 1.0618204 0.9645046 1.097237
#> 7  0.6 1.0777084 1.0548045 0.9393357 1.096963
#> 8  0.7 1.0741728 1.0479256 0.9149275 1.096690
#> 9  0.8 1.0706283 1.0411906 0.8913451 1.096416
#> 10 0.9 1.0670771 1.0346054 0.8686402 1.096143
#> 11 1.0 1.0635215 1.0281744 0.8468513 1.095871
#> 12 1.1 1.0599638 1.0219012 0.8260045 1.095599
#> 13 1.2 1.0564064 1.0157878 0.8061137 1.095327
#> 14 1.3 1.0528516 1.0098357 0.7871823 1.095056
#> 15 1.4 1.0493016 1.0040451 0.7692042 1.094785
#> 16 1.5 1.0457588 0.9984157 0.7521650 1.094515
#> 17 1.6 1.0422255 0.9929463 0.7360433 1.094245
#> 18 1.7 1.0387040 0.9876354 0.7208122 1.093976
#> 19 1.8 1.0351963 0.9824806 0.7064402 1.093707
#> 20 1.9 1.0317048 0.9774794 0.6928927 1.093439
#> 21 2.0 1.0282315 0.9726288 0.6801327 1.093172
#> 22 2.1 1.0247784 0.9679256 0.6681218 1.092905
#> 23 2.2 1.0213476 0.9633662 0.6568209 1.092639
#> 24 2.3 1.0179410 0.9589470 0.6461911 1.092373
#> 25 2.4 1.0145603 0.9546643 0.6361937 1.092109
#> 26 2.5 1.0112074 0.9505141 0.6267911 1.091844
#> 27 2.6 1.0078840 0.9464927 0.6179467 1.091581
#> 28 2.7 1.0045915 0.9425961 0.6096255 1.091318
#> 29 2.8 1.0013316 0.9388203 0.6017940 1.091056
#> 30 2.9 0.9981055 0.9351616 0.5944204 1.090795
#> 31 3.0 0.9949147 0.9316160 0.5874745 1.090535
#> 
#> attr(,"R")
#> [1] 100
#> attr(,"conf")
#> [1] 0.95
#> attr(,"parameter")
#> [1] "renyi"
#> attr(,"ci.type")
#> [1] "perc"
#> 
#> $`Diversity profiles`$tsallis
#> $`Diversity profiles`$tsallis$Long
#>      q  observed      mean     lower     upper
#> 1  0.0 3.0000000 2.9900000 3.0000000 3.0000000
#> 2  0.1 2.6589612 2.6378910 2.4996223 2.7161679
#> 3  0.2 2.3713503 2.3447960 2.1181515 2.4676103
#> 4  0.3 2.1274272 2.0988213 1.8235216 2.2492716
#> 5  0.4 1.9193870 1.8907700 1.5928772 2.0568872
#> 6  0.5 1.7409539 1.7134770 1.4098232 1.8868521
#> 7  0.6 1.5870644 1.5613196 1.2625179 1.7361131
#> 8  0.7 1.4536179 1.4298533 1.1567171 1.6020787
#> 9  0.8 1.3372809 1.3155399 1.0650894 1.4825443
#> 10 0.9 1.2353333 1.2155430 0.9801273 1.3756306
#> 11 1.0 1.1455462 1.1275734 0.9078287 1.2797376
#> 12 1.1 1.0660859 1.0497713 0.8455749 1.1934717
#> 13 1.2 0.9954373 0.9806158 0.7914366 1.1156705
#> 14 1.3 0.9323437 0.9188559 0.7416317 1.0453117
#> 15 1.4 0.8757575 0.8634558 0.6948633 0.9815190
#> 16 1.5 0.8248023 0.8135534 0.6534720 0.9235343
#> 17 1.6 0.7787414 0.7684270 0.6166461 0.8707008
#> 18 1.7 0.7369530 0.7274689 0.5837189 0.8224474
#> 19 1.8 0.6989100 0.6901654 0.5541388 0.7782770
#> 20 1.9 0.6641637 0.6560795 0.5274471 0.7377551
#> 21 2.0 0.6323302 0.6248380 0.5032600 0.7005015
#> 22 2.1 0.6030805 0.5961205 0.4812547 0.6661827
#> 23 2.2 0.5761307 0.5696510 0.4611581 0.6345670
#> 24 2.3 0.5512352 0.5451907 0.4427386 0.6054311
#> 25 2.4 0.5281808 0.5225320 0.4257981 0.5784096
#> 26 2.5 0.5067815 0.5014938 0.4101673 0.5533067
#> 27 2.6 0.4868749 0.4819176 0.3957002 0.5299479
#> 28 2.7 0.4683181 0.4636643 0.3822711 0.5081781
#> 29 2.8 0.4509853 0.4466111 0.3697706 0.4878587
#> 30 2.9 0.4347655 0.4306495 0.3581038 0.4688654
#> 31 3.0 0.4195602 0.4156832 0.3471879 0.4510869
#> 
#> $`Diversity profiles`$tsallis$Medium
#>      q  observed      mean     lower     upper
#> 1  0.0 3.0000000 3.0000000 3.0000000 3.0000000
#> 2  0.1 2.6732473 2.6523415 2.5743673 2.7140085
#> 3  0.2 2.3966283 2.3636721 2.2412466 2.4639346
#> 4  0.3 2.1610045 2.1217618 1.9751598 2.2445591
#> 5  0.4 1.9590700 1.9172331 1.7635515 2.0518492
#> 6  0.5 1.7849628 1.7428433 1.5832919 1.8820998
#> 7  0.6 1.6339616 1.5929579 1.4319269 1.7318509
#> 8  0.7 1.5022489 1.4631610 1.3068920 1.5984064
#> 9  0.8 1.3867251 1.3499657 1.1989887 1.4794915
#> 10 0.9 1.2848625 1.2505983 1.1022878 1.3731815
#> 11 1.0 1.1945906 1.1628361 1.0177378 1.2778429
#> 12 1.1 1.1142058 1.0848843 0.9445421 1.1920851
#> 13 1.2 1.0422993 1.0152838 0.8805382 1.1147205
#> 14 1.3 0.9777011 0.9528392 0.8225672 1.0447279
#> 15 1.4 0.9194342 0.8965647 0.7698358 0.9812134
#> 16 1.5 0.8666793 0.8456416 0.7229659 0.9237489
#> 17 1.6 0.8187454 0.7993857 0.6811250 0.8713677
#> 18 1.7 0.7750476 0.7572214 0.6436161 0.8234740
#> 19 1.8 0.7350876 0.7186616 0.6098537 0.7795816
#> 20 1.9 0.6984396 0.6832915 0.5793439 0.7392671
#> 21 2.0 0.6647377 0.6507562 0.5516686 0.7021605
#> 22 2.1 0.6336659 0.6207502 0.5264728 0.6679337
#> 23 2.2 0.6049505 0.5930093 0.5034538 0.6363063
#> 24 2.3 0.5783532 0.5673040 0.4823529 0.6070271
#> 25 2.4 0.5536656 0.5434338 0.4629479 0.5798739
#> 26 2.5 0.5307051 0.5212231 0.4450475 0.5546499
#> 27 2.6 0.5093107 0.5005173 0.4284866 0.5311799
#> 28 2.7 0.4893400 0.4811798 0.4131220 0.5093080
#> 29 2.8 0.4706669 0.4630896 0.3988293 0.4888948
#> 30 2.9 0.4531792 0.4461389 0.3854999 0.4698155
#> 31 3.0 0.4367766 0.4302317 0.3730390 0.4519579
#> 
#> $`Diversity profiles`$tsallis$Short
#>      q  observed      mean     lower     upper
#> 1  0.0 2.0000000 2.0000000 2.0000000 2.0000000
#> 2  0.1 1.8661938 1.8539385 1.7787768 1.8740070
#> 3  0.2 1.7437218 1.7223719 1.5950475 1.7577557
#> 4  0.3 1.6315129 1.6035690 1.4413061 1.6504216
#> 5  0.4 1.5286044 1.4960382 1.3116710 1.5512539
#> 6  0.5 1.4341299 1.3984902 1.2015178 1.4595689
#> 7  0.6 1.3473096 1.3098066 1.1071974 1.3747430
#> 8  0.7 1.2674411 1.2290140 1.0258192 1.2962082
#> 9  0.8 1.1938915 1.1552627 0.9550842 1.2234462
#> 10 0.9 1.1260900 1.0878088 0.8931564 1.1559841
#> 11 1.0 1.0635215 1.0259993 0.8368996 1.0933903
#> 12 1.1 1.0057210 0.9692595 0.7867645 1.0352708
#> 13 1.2 0.9522684 0.9170824 0.7419856 0.9812656
#> 14 1.3 0.9027839 0.8690195 0.7017921 0.9310457
#> 15 1.4 0.8569240 0.8246732 0.6655454 0.8843104
#> 16 1.5 0.8143777 0.7836906 0.6327139 0.8407845
#> 17 1.6 0.7748634 0.7457574 0.6057917 0.8002164
#> 18 1.7 0.7381257 0.7105933 0.5810880 0.7623757
#> 19 1.8 0.7039330 0.6779480 0.5582623 0.7270513
#> 20 1.9 0.6720751 0.6475973 0.5370839 0.6940500
#> 21 2.0 0.6423611 0.6193403 0.5173611 0.6631944
#> 22 2.1 0.6146175 0.5929964 0.4980090 0.6343222
#> 23 2.2 0.5886863 0.5684033 0.4799675 0.6072842
#> 24 2.3 0.5644238 0.5454146 0.4631037 0.5819435
#> 25 2.4 0.5416992 0.5238983 0.4473029 0.5581744
#> 26 2.5 0.5203930 0.5037350 0.4324653 0.5358616
#> 27 2.6 0.5003964 0.4848167 0.4179304 0.5148987
#> 28 2.7 0.4816099 0.4670456 0.4040632 0.4951883
#> 29 2.8 0.4639427 0.4503326 0.3910486 0.4766405
#> 30 2.9 0.4473116 0.4345971 0.3788115 0.4591725
#> 31 3.0 0.4316406 0.4197656 0.3672852 0.4427083
#> 
#> attr(,"R")
#> [1] 100
#> attr(,"conf")
#> [1] 0.95
#> attr(,"parameter")
#> [1] "tsallis"
#> attr(,"ci.type")
#> [1] "perc"
#> 
#>