Parse Output Files from MStrat

Prepare raw output files generated by MStrat (Schoen and Brown 1993; Gouesnard et al. 2001; Gouesnard et al. 2002) .

Usage

parse_mstrat_out(
  data.file,
  genotype = NULL,
  variable.file,
  kernel.file,
  redundance.output = NULL,
  core.output = NULL
)

Arguments

data.file

The path to the .dat input file used for generating the output files from MStrat.

genotype

Name of column/variable with the genotype names as a character string.

variable.file

The path to the .var input file used for generating the output files from MStrat.

kernel.file

The path to the .ker input file used for generating the output files from MStrat.

redundance.output

The path to the redundance output file generated by MStrat.

core.output

The path to the core output file generated by MStrat.

Value

A list with the following components:

MStrat Core Output

The raw output of core sets constructed by MStrat.

MStrat Core Optimised

The optimised core set output from MStrat.

MStrat Redundance Output

The raw output of Redundance estimation by MStrat

MStrat Redundance Plot

A list of plots of Redundance estimation.

References

Gouesnard B, Bataillon TM, Decoux G, Rozale C, Schoen DJ, David JL (2001). “MSTRAT: An algorithm for building germ plasm core collections by maximizing allelic or phenotypic richness.” Journal of Heredity, 92(1), 93–94.

Gouesnard B, Bataillon TM, Decoux G, Rozale C, Schoen DJ, David JL (2002). “MStrat Documentation 1.1.” Evolutionary genomics and population management (GE\(^{\textrm{2}}\)pop), Institut Amelioration Genetique et Adaptation des Plantes mediterraneennes et tropicales (agAp Institute), CIRAD, Montpellier, France.

Schoen DJ, Brown AH (1993). “Conservation of allelic richness in wild crop relatives is aided by assessment of genetic markers.” Proceedings of the National Academy of Sciences, 90(22), 10623–10627.

See also

Examples

if (FALSE) { # interactive()

parse_mstrat_out(data.file = "MStrat_input_data.dat",
                 genotype = "Accession",
                 variable.file = "MStrat_input_variable.var",
                 kernel.file = "MStrat_input_kernel.ker",
                 redundance.output = "MStrat - Redundance.out",
                 core.output = "MStrat - Core.out")
}