Compare Diversity Measures

Compare Diversity Measures

Usage

diversity.compare(
  x,
  group,
  R = 1000,
  base = exp(1),
  na.omit = TRUE,
  global.test = TRUE,
  pairwise.test = TRUE,
  bootstrap.ci = TRUE,
  diversity.profile = TRUE,
  p.adjust.method = c("bonferroni", "holm"),
  ci.conf = 0.95,
  ci.type = c("perc", "bca"),
  q = seq(0, 3, 0.1),
  parallel = c("no", "multicore", "snow"),
  ncpus = 1L,
  cl = NULL
)

Arguments

x

A factor vector of categories (e.g., species, traits). The frequency of each level is treated as the abundance of that category.

group

A factor vector indicating the group of each observation. Must have the same length as x.

R

Integer specifying the number of permutations. Default is 1000.

base

The logarithm base to be used for computation of shannon family of diversity indices. Default is exp(1).

na.omit

logical. If TRUE, missing values (NA) are ignored and not included as a distinct factor level for computation. Default is TRUE.

global.test

logical. If TRUE performs the global permutation tests for the diversity measures. Default is TRUE.

pairwise.test

logical. If TRUE performs the pairwise permutation tests for the diversity measures. Default is TRUE.

bootstrap.ci

logical. If TRUE computes the bootstrap confidence intervals for the diversity measures. Default is TRUE.

diversity.profile

logical. If TRUE diversity profiles. Default is TRUE.

p.adjust.method

(perm.test.pairwise only) Method for adjusting p-values for multiple comparisons. Options include "bonferroni" and "holm". Default is "bonferroni".

ci.conf

Confidence level of the bootstrap interval. Default is 0.95.

ci.type

A vector of character strings representing the type of intervals required. The options are c("perc", "bca").

q

The order of the parametric index.

parallel

The type of parallel operation to be used (if any). If missing, the default is taken from the option "boot.parallel" (and if that is not set, "no").

ncpus

integer: number of processes to be used in parallel operation: typically one would chose this to the number of available CPUs.

cl

An optional parallel or snow cluster for use if parallel = "snow". If not supplied, a cluster on the local machine is created for the duration of the boot call.

Value

A list with the following elements.

Diversity Indices

A data frame of the different diversity indices computed for each group.

Global Test

A data frame of results of global permutation test including the test statistic (weighted sum of squares between group summary indices) and the p value for the different diversity indices.

Pairwise Test

A list of the following data frames.

p-value

A data frame of p values for each between group comparison for different diversity measures.

cld

A data frame of compact letter displays of significant differences among groups for different diversity measures.

Bootstrap CIs

A data frame of lower and upper bootstrap confidence intervals computed for each group in different diversity measures.

Diversity profiles

A list of data frames of Hill, Renyi and Tsallis diversity profiles computed for each group.

Examples

library(EvaluateCore)

pdata <- cassava_CC

qual <- c("CUAL", "LNGS", "PTLC", "DSTA", "LFRT", "LBTEF", "CBTR", "NMLB",
          "ANGB", "CUAL9M", "LVC9M", "TNPR9M", "PL9M", "STRP", "STRC",
          "PSTR")

# Convert qualitative data columns to factor
pdata[, qual] <- lapply(pdata[, qual], as.factor)

str(pdata)
#> 'data.frame':	168 obs. of  26 variables:
#>  $ CUAL  : Factor w/ 4 levels "Dark green","Green purple",..: 3 1 2 2 2 2 4 2 2 1 ...
#>  $ LNGS  : Factor w/ 3 levels "Long","Medium",..: 3 1 2 2 2 2 2 1 1 1 ...
#>  $ PTLC  : Factor w/ 5 levels "Dark green","Green purple",..: 3 4 4 4 4 5 4 2 2 5 ...
#>  $ DSTA  : Factor w/ 5 levels "Absent","Central part",..: 1 5 5 5 5 5 5 4 2 5 ...
#>  $ LFRT  : Factor w/ 4 levels "25-50% leaf retention",..: 1 1 1 1 3 2 2 2 2 2 ...
#>  $ LBTEF : Factor w/ 6 levels "0","1","2","3",..: 3 1 2 1 4 5 4 4 3 2 ...
#>  $ CBTR  : Factor w/ 3 levels "Cream","White",..: 2 2 2 2 1 2 1 1 1 1 ...
#>  $ NMLB  : Factor w/ 9 levels "0","1","2","3",..: 3 1 2 1 4 4 4 3 3 4 ...
#>  $ ANGB  : Factor w/ 4 levels "150-300","450-600",..: 1 4 1 4 2 2 2 1 2 2 ...
#>  $ CUAL9M: Factor w/ 5 levels "Dark green","Green",..: 1 1 3 5 3 3 5 5 5 4 ...
#>  $ LVC9M : Factor w/ 5 levels "Dark green","Green",..: 4 3 3 3 3 1 3 1 4 3 ...
#>  $ TNPR9M: Factor w/ 5 levels "1","2","3","4",..: 5 5 4 2 5 4 2 5 5 5 ...
#>  $ PL9M  : Factor w/ 2 levels "Long (25-30cm)",..: 2 2 1 1 1 1 1 1 2 2 ...
#>  $ STRP  : Factor w/ 4 levels "Absent","Intermediate",..: 2 3 1 1 1 1 4 1 1 4 ...
#>  $ STRC  : Factor w/ 2 levels "Absent","Present": 2 2 1 2 1 1 2 1 1 2 ...
#>  $ PSTR  : Factor w/ 2 levels "Irregular","Tending toward horizontal": 1 2 2 2 1 2 2 2 1 2 ...
#>  $ NMSR  : num  6 2 6 2 20 13 4 14 10 5 ...
#>  $ TTRN  : num  3 0.5 3 2 5 ...
#>  $ TFWSR : num  1.4 2.6 1.2 1.6 5 7 4.2 2.8 2.8 4 ...
#>  $ TTRW  : num  0.7 0.65 0.6 1.6 1.25 ...
#>  $ TFWSS : num  1 2.8 2.8 2.4 16 12 9 4.4 6.2 5 ...
#>  $ TTSW  : num  0.5 0.7 1.4 2.4 4 ...
#>  $ TTPW  : num  2.4 5.4 4 4 21 19 13.2 7.2 9 9 ...
#>  $ AVPW  : num  1.2 1.35 2 4 5.25 4.75 3.3 2.4 1.8 2.25 ...
#>  $ ARSR  : num  2 0 2 0 3 0 0 6 0 0 ...
#>  $ SRDM  : num  42 39.8 29.7 43 37.9 37 38.9 36.9 41 37.9 ...

diversity.compare(x = pdata$CUAL, group = pdata$LNGS, R = 100,
                  base = exp(1), na.omit = TRUE)
#> Computing diversity indices.
#> Performing global permutation tests.
#> 
#> Performing pairwise permutation tests.
#> Computing bootstrap confidence intervals.
#> Generating diversity profiles.
#> $`Diversity Indices`
#> $`Diversity Indices`$Indices
#> # A tibble: 4 × 30
#>   group   richness margalef_index menhinick_index berger_parker
#>   <chr>      <int>          <dbl>           <dbl>         <dbl>
#> 1 Overall        4          0.585           0.309         0.530
#> 2 Long           3          0.468           0.354         0.514
#> 3 Medium         4          0.701           0.471         0.569
#> 4 Short          4          0.944           0.816         0.458
#> # ℹ 25 more variables: berger_parker_reciprocal <dbl>, simpson <dbl>,
#> #   gini_simpson <dbl>, simpson_max <dbl>, simpson_reciprocal <dbl>,
#> #   simpson_relative <dbl>, simpson_evenness <dbl>, shannon <dbl>,
#> #   shannon_max <dbl>, shannon_relative <dbl>, shannon_ens <dbl>,
#> #   heip_evenness <dbl>, mcintosh_diversity <dbl>, mcintosh_evenness <dbl>,
#> #   smith_wilson <dbl>, brillouin_index <dbl>, renyi_entropy_0 <dbl>,
#> #   renyi_entropy_1 <dbl>, renyi_entropy_2 <dbl>, tsallis_entropy_0 <dbl>, …
#> 
#> $`Diversity Indices`$Warnings
#> $`Diversity Indices`$Warnings$Overall
#> character(0)
#> 
#> $`Diversity Indices`$Warnings$Long
#> character(0)
#> 
#> $`Diversity Indices`$Warnings$Medium
#> character(0)
#> 
#> $`Diversity Indices`$Warnings$Short
#> character(0)
#> 
#> 
#> 
#> $`Global Test`
#>          Measure margalef_index menhinick_index berger_parker
#> 1 Test statistic     4.62557535      3.85787925     0.2539683
#> 2        p-value     0.06930693      0.06930693     0.6831683
#>   berger_parker_reciprocal   simpson gini_simpson simpson_max simpson_relative
#> 1                3.5485816 0.1459742    0.1459742  0.28571429        0.6397893
#> 2                0.6039604 0.5544554    0.5544554  0.08910891        0.1782178
#>     shannon shannon_max shannon_relative shannon_ens heip_evenness
#> 1 2.1255498  7.08710918       1.15089028   9.7353113   10.23058958
#> 2 0.3168317  0.08910891       0.02970297   0.2673267    0.03960396
#>   mcintosh_diversity mcintosh_evenness smith_wilson brillouin_index
#> 1          0.3514116         1.0809743   3.78072900       0.2103725
#> 2          0.2376238         0.1485149   0.01980198       0.7821782
#> 
#> $`Pairwise Test`
#> $`Pairwise Test`$`p-value`
#>        Comparison margalef_index menhinick_index berger_parker
#> 1  Long vs Medium     1.00000000      1.00000000             1
#> 2   Long vs Short     0.02970297      0.02970297             1
#> 3 Medium vs Short     1.00000000      1.00000000             1
#>   berger_parker_reciprocal   simpson gini_simpson simpson_max simpson_relative
#> 1                1.0000000 1.0000000    1.0000000   1.0000000        0.8019802
#> 2                1.0000000 1.0000000    1.0000000   0.6237624        1.0000000
#> 3                0.8910891 0.8316832    0.8316832   1.0000000        0.8019802
#>     shannon shannon_max shannon_relative shannon_ens heip_evenness
#> 1 1.0000000   1.0000000        0.9504950   1.0000000     0.9504950
#> 2 0.4158416   0.6237624        1.0000000   0.2970297     1.0000000
#> 3 0.6237624   1.0000000        0.2970297   0.5643564     0.2970297
#>   mcintosh_diversity mcintosh_evenness smith_wilson brillouin_index
#> 1          1.0000000         0.8019802   1.00000000               1
#> 2          0.2376238         1.0000000   1.00000000               1
#> 3          0.2376238         0.7425743   0.05940594               1
#> 
#> $`Pairwise Test`$cld
#>    Group margalef_index menhinick_index berger_parker berger_parker_reciprocal
#> 1   Long              a               a             a                        a
#> 2 Medium             ab              ab             a                        a
#> 3  Short              b               b             a                        a
#>   simpson gini_simpson simpson_max simpson_relative shannon shannon_max
#> 1       a            a           a                a       a           a
#> 2       a            a           a                a       a           a
#> 3       a            a           a                a       a           a
#>   shannon_relative shannon_ens heip_evenness mcintosh_diversity
#> 1                a           a             a                  a
#> 2                a           a             a                  a
#> 3                a           a             a                  a
#>   mcintosh_evenness smith_wilson brillouin_index
#> 1                 a            a               a
#> 2                 a            a               a
#> 3                 a            a               a
#> 
#> 
#> $`Bootstrap CIs`
#> $`Bootstrap CIs`$`Bootstrap CIs`
#>        Group-CI margalef_index menhinick_index berger_parker
#> 1   Long: lower      0.4676540       0.3535534     0.4100694
#> 2   Long: upper      0.4676540       0.3535534     0.6454861
#> 3 Medium: lower      0.4676540       0.3535534     0.4517361
#> 4 Medium: upper      0.7014810       0.4714045     0.6871528
#> 5  Short: lower      0.6293160       0.6123724     0.3333333
#> 6  Short: upper      0.9439739       0.8164966     0.7083333
#>   berger_parker_reciprocal   simpson gini_simpson simpson_max simpson_relative
#> 1                 1.549399 0.3465085    0.5192130   0.6666667        0.7788194
#> 2                 2.439310 0.4807870    0.6534915   0.6666667        0.9802373
#> 3                 1.455429 0.3511767    0.4750965   0.6666667        0.6530478
#> 4                 2.214205 0.5249035    0.6488233   0.7500000        0.9462674
#> 5                 1.411765 0.2743056    0.4431424   0.6666667        0.6276765
#> 6                 3.000000 0.5568576    0.7256944   0.7500000        0.9730903
#>    shannon shannon_max shannon_relative shannon_ens heip_evenness
#> 1 1.259241    1.584963        0.7944926    2.393756     1.2614621
#> 2 1.556697    1.584963        0.9821663    2.941797     1.8715668
#> 3 1.212174    1.584963        0.6454440    2.316935     0.8786718
#> 4 1.648590    2.000000        0.9536829    3.135288     1.7670811
#> 5 1.167107    1.584963        0.6203256    2.246508     0.8199156
#> 6 1.927082    2.000000        0.9735286    3.802881     1.9565144
#>   mcintosh_diversity mcintosh_evenness smith_wilson brillouin_index
#> 1          0.3475900         0.7254853    0.5004391       0.8172290
#> 2          0.4663058         0.9732672    0.8487120       1.0172578
#> 3          0.3123495         0.5714145    0.3524329       0.7781839
#> 4          0.4618275         0.9285291    0.7733972       1.0631896
#> 5          0.3188860         0.5461950    0.3987736       0.6618217
#> 6          0.5984076         0.9616910    0.8140394       1.1398823
#> 
#> $`Bootstrap CIs`$Warnings
#> # A tibble: 0 × 0
#> 
#> 
#> $`Diversity profiles`
#> $`Diversity profiles`$`Diversity profiles`
#> $`Diversity profiles`$`Diversity profiles`[[1]]
#> $`Diversity profiles`$`Diversity profiles`[[1]]$Long
#>      q observed     mean    lower    upper ci.type
#> 1  0.0 3.000000 3.000000 3.000000 3.000000    perc
#> 2  0.1 2.970041 2.965527 2.919290 2.994121    perc
#> 3  0.2 2.940750 2.932192 2.842233 2.988250    perc
#> 4  0.3 2.912153 2.900010 2.769187 2.982390    perc
#> 5  0.4 2.884272 2.868987 2.700099 2.976541    perc
#> 6  0.5 2.857124 2.839120 2.634955 2.970706    perc
#> 7  0.6 2.830721 2.810399 2.573706 2.964887    perc
#> 8  0.7 2.805073 2.782809 2.516270 2.959084    perc
#> 9  0.8 2.780186 2.756329 2.469493 2.953301    perc
#> 10 0.9 2.756060 2.730934 2.429612 2.947538    perc
#> 11 1.0 2.732695 2.706596 2.393756 2.941797    perc
#> 12 1.1 2.710085 2.683284 2.361469 2.936080    perc
#> 13 1.2 2.688224 2.660965 2.326622 2.930389    perc
#> 14 1.3 2.667101 2.639605 2.295192 2.924724    perc
#> 15 1.4 2.646704 2.619169 2.264013 2.919089    perc
#> 16 1.5 2.627019 2.599620 2.230264 2.913484    perc
#> 17 1.6 2.608031 2.580922 2.197525 2.907910    perc
#> 18 1.7 2.589722 2.563040 2.166801 2.902370    perc
#> 19 1.8 2.572075 2.545938 2.136294 2.896865    perc
#> 20 1.9 2.555070 2.529581 2.107384 2.891396    perc
#> 21 2.0 2.538688 2.513935 2.080265 2.885964    perc
#> 22 2.1 2.522908 2.498966 2.054843 2.880572    perc
#> 23 2.2 2.507711 2.484642 2.031026 2.875219    perc
#> 24 2.3 2.493075 2.470933 2.008719 2.869908    perc
#> 25 2.4 2.478981 2.457808 1.987832 2.864639    perc
#> 26 2.5 2.465409 2.445239 1.968274 2.859414    perc
#> 27 2.6 2.452337 2.433199 1.949959 2.854234    perc
#> 28 2.7 2.439748 2.421660 1.932804 2.849100    perc
#> 29 2.8 2.427621 2.410598 1.916731 2.844013    perc
#> 30 2.9 2.415938 2.399990 1.901666 2.838973    perc
#> 31 3.0 2.404680 2.389812 1.887538 2.833982    perc
#> 
#> $`Diversity profiles`$`Diversity profiles`[[1]]$Medium
#>      q observed     mean    lower    upper ci.type
#> 1  0.0 4.000000 3.540000 3.000000 4.000000    perc
#> 2  0.1 3.775167 3.412327 2.913982 3.869732    perc
#> 3  0.2 3.586424 3.299979 2.832172 3.749020    perc
#> 4  0.3 3.427791 3.200860 2.756875 3.638839    perc
#> 5  0.4 3.293906 3.113045 2.687079 3.537879    perc
#> 6  0.5 3.180150 3.034831 2.622537 3.450294    perc
#> 7  0.6 3.082659 2.964754 2.563273 3.373527    perc
#> 8  0.7 2.998274 2.901581 2.504153 3.304748    perc
#> 9  0.8 2.924454 2.844287 2.445643 3.242893    perc
#> 10 0.9 2.859178 2.792026 2.383772 3.186604    perc
#> 11 1.0 2.800853 2.847501 2.353253 3.189054    perc
#> 12 1.1 2.748230 2.699954 2.256499 3.088438    perc
#> 13 1.2 2.700330 2.659105 2.201718 3.045469    perc
#> 14 1.3 2.656390 2.621170 2.151954 3.005878    perc
#> 15 1.4 2.615811 2.585830 2.106659 2.966002    perc
#> 16 1.5 2.578125 2.552814 2.065359 2.937614    perc
#> 17 1.6 2.542961 2.521895 2.027642 2.912390    perc
#> 18 1.7 2.510025 2.492881 1.993147 2.898757    perc
#> 19 1.8 2.479080 2.465605 1.961553 2.883340    perc
#> 20 1.9 2.449934 2.439921 1.932579 2.865014    perc
#> 21 2.0 2.422430 2.415705 1.905974 2.847655    perc
#> 22 2.1 2.396434 2.392844 1.881511 2.831142    perc
#> 23 2.2 2.371835 2.371239 1.858991 2.815373    perc
#> 24 2.3 2.348536 2.350800 1.838233 2.800269    perc
#> 25 2.4 2.326451 2.331447 1.819073 2.785760    perc
#> 26 2.5 2.305505 2.313107 1.801367 2.771792    perc
#> 27 2.6 2.285629 2.295713 1.784984 2.758318    perc
#> 28 2.7 2.266761 2.279205 1.769803 2.745290    perc
#> 29 2.8 2.248843 2.263526 1.755720 2.732512    perc
#> 30 2.9 2.231821 2.248624 1.742637 2.720138    perc
#> 31 3.0 2.215647 2.234452 1.730468 2.708146    perc
#> 
#> $`Diversity profiles`$`Diversity profiles`[[1]]$Short
#>      q observed     mean    lower    upper ci.type
#> 1  0.0 4.000000 3.850000 3.000000 4.000000    perc
#> 2  0.1 3.933810 3.757118 2.928167 3.978361    perc
#> 3  0.2 3.870196 3.669992 2.857629 3.957155    perc
#> 4  0.3 3.809136 3.588497 2.788678 3.936381    perc
#> 5  0.4 3.750595 3.512442 2.721584 3.916039    perc
#> 6  0.5 3.694521 3.441585 2.656591 3.896127    perc
#> 7  0.6 3.640856 3.375645 2.593911 3.876641    perc
#> 8  0.7 3.589527 3.314320 2.519231 3.857579    perc
#> 9  0.8 3.540459 3.257302 2.427858 3.838935    perc
#> 10 0.9 3.493567 3.204280 2.340655 3.820704    perc
#> 11 1.0 3.448767 3.154954 2.246508 3.802881    perc
#> 12 1.1 3.405970 3.109039 2.164395 3.785459    perc
#> 13 1.2 3.365087 3.066263 2.092723 3.768432    perc
#> 14 1.3 3.326031 3.026377 2.030060 3.751792    perc
#> 15 1.4 3.288713 2.989147 1.977788 3.735533    perc
#> 16 1.5 3.253050 2.954359 1.938795 3.719648    perc
#> 17 1.6 3.218958 2.921817 1.903960 3.704129    perc
#> 18 1.7 3.186359 2.891341 1.872727 3.688969    perc
#> 19 1.8 3.155176 2.862767 1.844620 3.674161    perc
#> 20 1.9 3.125338 2.835945 1.819234 3.659697    perc
#> 21 2.0 3.096774 2.810738 1.796226 3.645570    perc
#> 22 2.1 3.069420 2.787022 1.775303 3.634102    perc
#> 23 2.2 3.043214 2.764682 1.756215 3.623070    perc
#> 24 2.3 3.018098 2.743615 1.738748 3.612454    perc
#> 25 2.4 2.994016 2.723726 1.722719 3.602234    perc
#> 26 2.5 2.970917 2.704928 1.707970 3.592392    perc
#> 27 2.6 2.948751 2.687142 1.694365 3.582911    perc
#> 28 2.7 2.927474 2.670296 1.681784 3.573773    perc
#> 29 2.8 2.907041 2.654323 1.670126 3.564963    perc
#> 30 2.9 2.887412 2.639162 1.659300 3.556464    perc
#> 31 3.0 2.868549 2.624758 1.649226 3.548262    perc
#> 
#> attr(,"R")
#> [1] 100
#> attr(,"conf")
#> [1] 0.95
#> attr(,"parameter")
#> [1] "hill"
#> attr(,"ci.type")
#> [1] "perc"
#> 
#> $`Diversity profiles`$`Diversity profiles`[[2]]
#> $`Diversity profiles`$`Diversity profiles`[[2]]$Long
#>      q  observed      mean     lower    upper ci.type
#> 1  0.0 1.0986123 1.0986123 1.0986123 1.098612    perc
#> 2  0.1 1.0885756 1.0870352 1.0713405 1.096651    perc
#> 3  0.2 1.0786646 1.0756751 1.0445900 1.094688    perc
#> 4  0.3 1.0688928 1.0645524 1.0185534 1.092725    perc
#> 5  0.4 1.0592726 1.0536840 0.9932854 1.090762    perc
#> 6  0.5 1.0498155 1.0430838 0.9688562 1.088800    perc
#> 7  0.6 1.0405315 1.0327627 0.9453233 1.086839    perc
#> 8  0.7 1.0314297 1.0227288 0.9227308 1.084880    perc
#> 9  0.8 1.0225177 1.0129878 0.9039650 1.082923    perc
#> 10 0.9 1.0138021 1.0035431 0.8876954 1.080970    perc
#> 11 1.0 1.0052882 0.9943962 0.8728393 1.079020    perc
#> 12 1.1 0.9969801 0.9855467 0.8592705 1.077075    perc
#> 13 1.2 0.9888807 0.9769926 0.8444168 1.075134    perc
#> 14 1.3 0.9809920 0.9687308 0.8308097 1.073199    perc
#> 15 1.4 0.9733151 0.9607569 0.8171116 1.071270    perc
#> 16 1.5 0.9658499 0.9530654 0.8020777 1.069347    perc
#> 17 1.6 0.9585956 0.9456502 0.7872769 1.067432    perc
#> 18 1.7 0.9515507 0.9385045 0.7731827 1.065525    perc
#> 19 1.8 0.9447130 0.9316209 0.7589977 1.063625    perc
#> 20 1.9 0.9380796 0.9249916 0.7453687 1.061735    perc
#> 21 2.0 0.9316472 0.9186086 0.7324130 1.059854    perc
#> 22 2.1 0.9254122 0.9124637 0.7201144 1.057982    perc
#> 23 2.2 0.9193702 0.9065485 0.7084532 1.056121    perc
#> 24 2.3 0.9135169 0.9008547 0.6974074 1.054271    perc
#> 25 2.4 0.9078477 0.8953740 0.6869526 1.052432    perc
#> 26 2.5 0.9023576 0.8900981 0.6770634 1.050605    perc
#> 27 2.6 0.8970416 0.8850190 0.6677135 1.048790    perc
#> 28 2.7 0.8918947 0.8801286 0.6588763 1.046988    perc
#> 29 2.8 0.8869118 0.8754193 0.6505251 1.045199    perc
#> 30 2.9 0.8820876 0.8708834 0.6426336 1.043423    perc
#> 31 3.0 0.8774170 0.8665138 0.6351764 1.041662    perc
#> 
#> $`Diversity profiles`$`Diversity profiles`[[2]]$Medium
#>      q  observed      mean     lower     upper ci.type
#> 1  0.0 1.3862944 1.2539606 1.0986123 1.3862944    perc
#> 2  0.1 1.3284446 1.2196338 1.0695199 1.3531850    perc
#> 3  0.2 1.2771555 1.1879051 1.0410417 1.3214922    perc
#> 4  0.3 1.2319160 1.1586511 1.0140956 1.2916590    perc
#> 5  0.4 1.1920742 1.1316895 0.9884536 1.2635166    perc
#> 6  0.5 1.1569284 1.1068094 0.9641420 1.2384492    perc
#> 7  0.6 1.1257925 1.0837948 0.9412842 1.2159523    perc
#> 8  0.7 1.0980367 1.0624396 0.9179494 1.1953579    perc
#> 9  0.8 1.0731077 1.0425563 0.8943079 1.1764658    perc
#> 10 0.9 1.0505342 1.0239795 0.8686818 1.1589550    perc
#> 11 1.0 1.0299241 1.0440692 0.8553457 1.1596702    perc
#> 12 1.1 1.0109570 0.9901955 0.8137426 1.1276507    perc
#> 13 1.2 0.9933740 0.9747635 0.7891224 1.1136260    perc
#> 14 1.3 0.9769680 0.9601836 0.7662219 1.1005218    perc
#> 15 1.4 0.9615743 0.9463820 0.7449180 1.0871727    perc
#> 16 1.5 0.9470625 0.9332962 0.7250965 1.0775935    perc
#> 17 1.6 0.9333293 0.9208723 0.7066515 1.0689660    perc
#> 18 1.7 0.9202928 0.9090639 0.6894846 1.0642578    perc
#> 19 1.8 0.9078876 0.8978302 0.6735042 1.0589221    perc
#> 20 1.9 0.8960613 0.8871351 0.6586249 1.0525525    perc
#> 21 2.0 0.8847711 0.8769463 0.6447671 1.0464808    perc
#> 22 2.1 0.8739819 0.8672344 0.6318562 1.0406695    perc
#> 23 2.2 0.8636641 0.8579729 0.6198228 1.0350879    perc
#> 24 2.3 0.8537922 0.8491371 0.6086022 1.0297111    perc
#> 25 2.4 0.8443440 0.8407042 0.5981339 1.0245184    perc
#> 26 2.5 0.8352998 0.8326530 0.5883618 1.0194929    perc
#> 27 2.6 0.8266413 0.8249635 0.5792337 1.0146207    perc
#> 28 2.7 0.8183519 0.8176171 0.5707014 1.0098866    perc
#> 29 2.8 0.8104158 0.8105961 0.5627199 1.0052210    perc
#> 30 2.9 0.8028180 0.8038840 0.5552479 1.0006819    perc
#> 31 3.0 0.7955444 0.7974649 0.5482469 0.9962622    perc
#> 
#> $`Diversity profiles`$`Diversity profiles`[[2]]$Short
#>      q observed      mean     lower    upper ci.type
#> 1  0.0 1.386294 1.3431421 1.0986123 1.386294    perc
#> 2  0.1 1.369609 1.3190363 1.0743761 1.380870    perc
#> 3  0.2 1.353305 1.2956683 1.0499905 1.375524    perc
#> 4  0.3 1.337402 1.2731155 1.0255639 1.370259    perc
#> 5  0.4 1.321914 1.2514372 1.0012082 1.365077    perc
#> 6  0.5 1.306851 1.2306731 0.9770358 1.359978    perc
#> 7  0.6 1.292219 1.2108452 0.9531570 1.354963    perc
#> 8  0.7 1.278021 1.1919591 0.9239536 1.350033    perc
#> 9  0.8 1.264256 1.1740066 0.8868933 1.345188    perc
#> 10 0.9 1.250923 1.1569682 0.8500964 1.340427    perc
#> 11 1.0 1.238017 1.1408164 0.8089771 1.335752    perc
#> 12 1.1 1.225530 1.1255173 0.7716779 1.331160    perc
#> 13 1.2 1.213454 1.1110333 0.7379438 1.326652    perc
#> 14 1.3 1.201780 1.0973247 0.7074902 1.322228    perc
#> 15 1.4 1.190496 1.0843509 0.6814069 1.317886    perc
#> 16 1.5 1.179593 1.0720714 0.6616248 1.313625    perc
#> 17 1.6 1.169058 1.0604463 0.6435955 1.309446    perc
#> 18 1.7 1.158879 1.0494374 0.6271330 1.305345    perc
#> 19 1.8 1.149044 1.0390077 0.6120705 1.301324    perc
#> 20 1.9 1.139542 1.0291220 0.5982590 1.297380    perc
#> 21 2.0 1.130361 1.0197471 0.5855667 1.293513    perc
#> 22 2.1 1.121489 1.0108515 0.5738773 1.290362    perc
#> 23 2.2 1.112914 1.0024057 0.5630882 1.287322    perc
#> 24 2.3 1.104627 0.9943817 0.5531090 1.284387    perc
#> 25 2.4 1.096616 0.9867537 0.5438602 1.281554    perc
#> 26 2.5 1.088871 0.9794972 0.5352717 1.278818    perc
#> 27 2.6 1.081382 0.9725895 0.5272815 1.276176    perc
#> 28 2.7 1.074140 0.9660093 0.5198349 1.273622    perc
#> 29 2.8 1.067136 0.9597367 0.5128832 1.271154    perc
#> 30 2.9 1.060361 0.9537532 0.5063832 1.268767    perc
#> 31 3.0 1.053806 0.9480415 0.5002964 1.266458    perc
#> 
#> attr(,"R")
#> [1] 100
#> attr(,"conf")
#> [1] 0.95
#> attr(,"parameter")
#> [1] "renyi"
#> attr(,"ci.type")
#> [1] "perc"
#> 
#> $`Diversity profiles`$`Diversity profiles`[[3]]
#> $`Diversity profiles`$`Diversity profiles`[[3]]$Long
#>      q  observed      mean     lower     upper ci.type
#> 1  0.0 2.0000000 2.0000000 2.0000000 2.0000000    perc
#> 2  0.1 1.8485612 1.8445078 1.8030064 1.8701492    perc
#> 3  0.2 1.7126236 1.7056892 1.6329542 1.7508452    perc
#> 4  0.3 1.5903508 1.5814316 1.4858245 1.6411364    perc
#> 5  0.4 1.4801432 1.4699205 1.3579911 1.5401630    perc
#> 6  0.5 1.3806058 1.3695933 1.2465007 1.4471473    perc
#> 7  0.6 1.2905205 1.2791022 1.1488930 1.3613862    perc
#> 8  0.7 1.2088223 1.1972819 1.0631117 1.2822429    perc
#> 9  0.8 1.1345790 1.1231228 0.9908472 1.2091410    perc
#> 10 0.9 1.0669734 1.0557484 0.9282917 1.1415579    perc
#> 11 1.0 1.0052882 0.9943962 0.8728393 1.0790200    perc
#> 12 1.1 0.9488928 0.9384013 0.8233871 1.0210976    perc
#> 13 1.2 0.8972324 0.8871830 0.7769629 0.9674007    perc
#> 14 1.3 0.8498176 0.8402328 0.7353630 0.9175750    perc
#> 15 1.4 0.8062167 0.7971048 0.6970026 0.8712985    perc
#> 16 1.5 0.7660477 0.7574069 0.6607378 0.8282785    perc
#> 17 1.6 0.7289728 0.7207937 0.6274374 0.7882486    perc
#> 18 1.7 0.6946920 0.6869605 0.5970673 0.7509663    perc
#> 19 1.8 0.6629392 0.6556377 0.5688731 0.7162107    perc
#> 20 1.9 0.6334774 0.6265866 0.5429826 0.6837803    perc
#> 21 2.0 0.6060957 0.5995949 0.5192130 0.6534915    perc
#> 22 2.1 0.5806057 0.5744741 0.4973384 0.6251766    perc
#> 23 2.2 0.5568392 0.5510560 0.4771604 0.5986823    perc
#> 24 2.3 0.5346455 0.5291904 0.4585041 0.5738689    perc
#> 25 2.4 0.5138895 0.5087429 0.4412154 0.5506082    perc
#> 26 2.5 0.4944499 0.4895931 0.4251580 0.5287832    perc
#> 27 2.6 0.4762176 0.4716328 0.4102118 0.5082866    perc
#> 28 2.7 0.4590941 0.4547646 0.3962702 0.4890202    perc
#> 29 2.8 0.4429909 0.4389010 0.3832386 0.4708937    perc
#> 30 2.9 0.4278278 0.4239626 0.3710331 0.4538246    perc
#> 31 3.0 0.4135320 0.4098779 0.3595792 0.4377369    perc
#> 
#> $`Diversity profiles`$`Diversity profiles`[[3]]$Medium
#>      q  observed      mean     lower     upper ci.type
#> 1  0.0 3.0000000 2.5400000 2.0000000 3.0000000    perc
#> 2  0.1 2.5617123 2.2401074 1.7982365 2.6444109    perc
#> 3  0.2 2.2224763 1.9956664 1.6247863 2.3478584    perc
#> 4  0.3 1.9552854 1.7936491 1.4767470 2.0998091    perc
#> 5  0.4 1.7411443 1.6245227 1.3492331 1.8904114    perc
#> 6  0.5 1.5665950 1.4812411 1.2388496 1.7149840    perc
#> 7  0.6 1.4220235 1.3585393 1.1429892 1.5660527    perc
#> 8  0.7 1.3004971 1.2524382 1.0567918 1.4377834    perc
#> 9  0.8 1.1969637 1.1598946 0.9792760 1.3263983    perc
#> 10 0.9 1.1076994 1.0785527 0.9075292 1.2287853    perc
#> 11 1.0 1.0299241 1.0440692 0.8553457 1.1596702    perc
#> 12 1.1 0.9615347 0.9424719 0.7815072 1.0663936    perc
#> 13 1.2 0.9009177 0.8850940 0.7299819 0.9983216    perc
#> 14 1.3 0.8468176 0.8334796 0.6844978 0.9372857    perc
#> 15 1.4 0.7982434 0.7868465 0.6441307 0.8816156    perc
#> 16 1.5 0.7544018 0.7445472 0.6081262 0.8330990    perc
#> 17 1.6 0.7146493 0.7060404 0.5758629 0.7890550    perc
#> 18 1.7 0.6784574 0.6708698 0.5468256 0.7503583    perc
#> 19 1.8 0.6453869 0.6386478 0.5205838 0.7141905    perc
#> 20 1.9 0.6150691 0.6090432 0.4967760 0.6802287    perc
#> 21 2.0 0.5871914 0.5817708 0.4750965 0.6488233    perc
#> 22 2.1 0.5614865 0.5565842 0.4552853 0.6197165    perc
#> 23 2.2 0.5377246 0.5332688 0.4371208 0.5926846    perc
#> 24 2.3 0.5157061 0.5116373 0.4204126 0.5675323    perc
#> 25 2.4 0.4952573 0.4915256 0.4049970 0.5440878    perc
#> 26 2.5 0.4762260 0.4727891 0.3907324 0.5221994    perc
#> 27 2.6 0.4584783 0.4553000 0.3774958 0.5017322    perc
#> 28 2.7 0.4418959 0.4389452 0.3651804 0.4825654    perc
#> 29 2.8 0.4263738 0.4236242 0.3536926 0.4645817    perc
#> 30 2.9 0.4118190 0.4092474 0.3429508 0.4476971    perc
#> 31 3.0 0.3981481 0.3957347 0.3328830 0.4318241    perc
#> 
#> $`Diversity profiles`$`Diversity profiles`[[3]]$Short
#>      q  observed      mean     lower     upper ci.type
#> 1  0.0 3.0000000 2.8500000 2.0000000 3.0000000    perc
#> 2  0.1 2.7003331 2.5444732 1.8109797 2.7391597    perc
#> 3  0.2 2.4405950 2.2846169 1.6454402 2.5067852    perc
#> 4  0.3 2.2146076 2.0623057 1.5001663 2.2993757    perc
#> 5  0.4 2.0172417 1.8710245 1.3724067 2.1138972    perc
#> 6  0.5 1.8442276 1.7055264 1.2598040 1.9477170    perc
#> 7  0.6 1.6920012 1.5615656 1.1603366 1.7985467    perc
#> 8  0.7 1.5575795 1.4356907 1.0647063 1.6643943    perc
#> 9  0.8 1.4384586 1.3250825 0.9704435 1.5435233    perc
#> 10 0.9 1.3325309 1.2274271 0.8873121 1.4344177    perc
#> 11 1.0 1.2380168 1.1408164 0.8089771 1.3357515    perc
#> 12 1.1 1.1534096 1.0636692 0.7426119 1.2463640    perc
#> 13 1.2 1.0774297 0.9946691 0.6859806 1.1652368    perc
#> 14 1.3 1.0089869 0.9327148 0.6373107 1.0914755    perc
#> 15 1.4 0.9471496 0.8768810 0.5962613 1.0242931    perc
#> 16 1.5 0.8911198 0.8263864 0.5631613 0.9629964    perc
#> 17 1.6 0.8402110 0.7805684 0.5337549 0.9069736    perc
#> 18 1.7 0.7938317 0.7388628 0.5074738 0.8556840    perc
#> 19 1.8 0.7514702 0.7007869 0.4838543 0.8086491    perc
#> 20 1.9 0.7126828 0.6659263 0.4625154 0.7654448    perc
#> 21 2.0 0.6770833 0.6339236 0.4431424 0.7256944    perc
#> 22 2.1 0.6443352 0.6044694 0.4254739 0.6892186    perc
#> 23 2.2 0.6141439 0.5772951 0.4092914 0.6555354    perc
#> 24 2.3 0.5862509 0.5521663 0.3944114 0.6243824    perc
#> 25 2.4 0.5604291 0.5288780 0.3806785 0.5955254    perc
#> 26 2.5 0.5364780 0.5072503 0.3679609 0.5687553    perc
#> 27 2.6 0.5142206 0.4871248 0.3561459 0.5438848    perc
#> 28 2.7 0.4934997 0.4683615 0.3451367 0.5207459    perc
#> 29 2.8 0.4741760 0.4508366 0.3348497 0.4991881    perc
#> 30 2.9 0.4561251 0.4344399 0.3252125 0.4790756    perc
#> 31 3.0 0.4392361 0.4190734 0.3161621 0.4602865    perc
#> 
#> attr(,"R")
#> [1] 100
#> attr(,"conf")
#> [1] 0.95
#> attr(,"parameter")
#> [1] "tsallis"
#> attr(,"ci.type")
#> [1] "perc"
#> 
#> 
#> $`Diversity profiles`$Warnings
#> $`Diversity profiles`$Warnings$hill
#> NULL
#> 
#> $`Diversity profiles`$Warnings$renyi
#> NULL
#> 
#> $`Diversity profiles`$Warnings$tsallis
#> NULL
#> 
#> 
#> 

diversity.compare(x = pdata$ANGB, group = pdata$LNGS, R = 100,
                  base = exp(1), na.omit = TRUE)
#> Computing diversity indices.
#> Performing global permutation tests.
#> 
#> Performing pairwise permutation tests.
#> Computing bootstrap confidence intervals.
#> Generating diversity profiles.
#> $`Diversity Indices`
#> $`Diversity Indices`$Indices
#> # A tibble: 4 × 30
#>   group   richness margalef_index menhinick_index berger_parker
#>   <chr>      <int>          <dbl>           <dbl>         <dbl>
#> 1 Overall        4          0.585           0.309         0.452
#> 2 Long           4          0.701           0.471         0.514
#> 3 Medium         4          0.701           0.471         0.444
#> 4 Short          3          0.629           0.612         0.458
#> # ℹ 25 more variables: berger_parker_reciprocal <dbl>, simpson <dbl>,
#> #   gini_simpson <dbl>, simpson_max <dbl>, simpson_reciprocal <dbl>,
#> #   simpson_relative <dbl>, simpson_evenness <dbl>, shannon <dbl>,
#> #   shannon_max <dbl>, shannon_relative <dbl>, shannon_ens <dbl>,
#> #   heip_evenness <dbl>, mcintosh_diversity <dbl>, mcintosh_evenness <dbl>,
#> #   smith_wilson <dbl>, brillouin_index <dbl>, renyi_entropy_0 <dbl>,
#> #   renyi_entropy_1 <dbl>, renyi_entropy_2 <dbl>, tsallis_entropy_0 <dbl>, …
#> 
#> $`Diversity Indices`$Warnings
#> $`Diversity Indices`$Warnings$Overall
#> character(0)
#> 
#> $`Diversity Indices`$Warnings$Long
#> character(0)
#> 
#> $`Diversity Indices`$Warnings$Medium
#> character(0)
#> 
#> $`Diversity Indices`$Warnings$Short
#> character(0)
#> 
#> 
#> 
#> $`Global Test`
#>          Measure margalef_index menhinick_index berger_parker
#> 1 Test statistic      0.1071317       0.4087945     0.1825397
#> 2        p-value      1.0000000       1.0000000     0.5940594
#>   berger_parker_reciprocal    simpson gini_simpson simpson_max simpson_relative
#> 1                3.4727768 0.03859249   0.03859249   0.1428571        0.2681430
#> 2                0.7524752 0.69306931   0.69306931   0.3861386        0.2376238
#>     shannon shannon_max shannon_relative shannon_ens heip_evenness
#> 1 0.6662977   3.5435546        0.3615327   3.0939334     3.0824569
#> 2 0.6336634   0.3861386        0.2673267   0.6534653     0.3960396
#>   mcintosh_diversity mcintosh_evenness smith_wilson brillouin_index
#> 1         0.07321862         0.4906392    1.6496292       0.6374954
#> 2         0.67326733         0.1881188    0.2277228       0.4752475
#> 
#> $`Pairwise Test`
#> $`Pairwise Test`$`p-value`
#>        Comparison margalef_index menhinick_index berger_parker
#> 1  Long vs Medium              1               1             1
#> 2   Long vs Short              1               1             1
#> 3 Medium vs Short              1               1             1
#>   berger_parker_reciprocal simpson gini_simpson simpson_max simpson_relative
#> 1                        1       1            1   1.0000000        1.0000000
#> 2                        1       1            1   0.8910891        0.4158416
#> 3                        1       1            1   1.0000000        0.6534653
#>   shannon shannon_max shannon_relative shannon_ens heip_evenness
#> 1       1   1.0000000        1.0000000           1     1.0000000
#> 2       1   0.8910891        0.3861386           1     0.6534653
#> 3       1   1.0000000        0.4752475           1     0.6831683
#>   mcintosh_diversity mcintosh_evenness smith_wilson brillouin_index
#> 1                  1         1.0000000    1.0000000               1
#> 2                  1         0.3267327    0.2079208               1
#> 3                  1         0.5940594    0.2970297               1
#> 
#> $`Pairwise Test`$cld
#>    Group margalef_index menhinick_index berger_parker berger_parker_reciprocal
#> 1   Long              a               a             a                        a
#> 2 Medium              a               a             a                        a
#> 3  Short              a               a             a                        a
#>   simpson gini_simpson simpson_max simpson_relative shannon shannon_max
#> 1       a            a           a                a       a           a
#> 2       a            a           a                a       a           a
#> 3       a            a           a                a       a           a
#>   shannon_relative shannon_ens heip_evenness mcintosh_diversity
#> 1                a           a             a                  a
#> 2                a           a             a                  a
#> 3                a           a             a                  a
#>   mcintosh_evenness smith_wilson brillouin_index
#> 1                 a            a               a
#> 2                 a            a               a
#> 3                 a            a               a
#> 
#> 
#> $`Bootstrap CIs`
#> $`Bootstrap CIs`$`Bootstrap CIs`
#>        Group-CI margalef_index menhinick_index berger_parker
#> 1   Long: lower       0.467654       0.3535534     0.4027778
#> 2   Long: upper       0.701481       0.4714045     0.6454861
#> 3 Medium: lower       0.701481       0.4714045     0.3684028
#> 4 Medium: upper       0.701481       0.4714045     0.5621528
#> 5  Short: lower       0.629316       0.6123724     0.3333333
#> 6  Short: upper       0.629316       0.6123724     0.6645833
#>   berger_parker_reciprocal   simpson gini_simpson simpson_max simpson_relative
#> 1                 1.549399 0.3121142    0.5264757   0.6666667        0.7019676
#> 2                 2.482759 0.4735243    0.6878858   0.7500000        0.9171811
#> 3                 1.779146 0.2924961    0.5859857   0.7500000        0.7813143
#> 4                 2.715385 0.4140143    0.7075039   0.7500000        0.9433385
#> 5                 1.510588 0.3333333    0.4850694   0.6666667        0.8023437
#> 6                 3.000000 0.5149306    0.6666667   0.6666667        1.0000000
#>    shannon shannon_max shannon_relative shannon_ens heip_evenness
#> 1 1.354753    1.584963        0.6869739    2.557567     0.9837284
#> 2 1.804062    2.000000        0.9055802    3.492233     1.6916805
#> 3 1.502925    2.000000        0.7514626    2.834168     1.1649393
#> 4 1.863664    2.000000        0.9318320    3.639311     1.8157760
#> 5 1.177499    1.584963        0.8284887    2.270659     1.3590567
#> 6 1.584963    1.584963        1.0000000    3.000000     1.9395542
#>   mcintosh_diversity mcintosh_evenness smith_wilson brillouin_index
#> 1          0.3535337         0.6237387    0.3509058       0.8742184
#> 2          0.5002876         0.8826564    0.6482790       1.1651868
#> 3          0.4042551         0.7132264    0.3537635       0.9685500
#> 4          0.5205152         0.9183439    0.6828628       1.2057380
#> 5          0.3558460         0.7524417    0.5050573       0.6959490
#> 6          0.5310498         1.0000000    1.0000000       0.9571217
#> 
#> $`Bootstrap CIs`$Warnings
#> # A tibble: 0 × 0
#> 
#> 
#> $`Diversity profiles`
#> $`Diversity profiles`$`Diversity profiles`
#> $`Diversity profiles`$`Diversity profiles`[[1]]
#> $`Diversity profiles`$`Diversity profiles`[[1]]$Long
#>      q observed     mean    lower    upper ci.type
#> 1  0.0 4.000000 3.970000 3.000000 4.000000    perc
#> 2  0.1 3.886394 3.835631 2.963291 3.941677    perc
#> 3  0.2 3.779642 3.712873 2.927296 3.885117    perc
#> 4  0.3 3.679590 3.600844 2.892063 3.830179    perc
#> 5  0.4 3.586026 3.498640 2.857637 3.776930    perc
#> 6  0.5 3.498684 3.405378 2.824056 3.725376    perc
#> 7  0.6 3.417265 3.320215 2.791353 3.675517    perc
#> 8  0.7 3.341444 3.242368 2.759555 3.627346    perc
#> 9  0.8 3.270881 3.171114 2.704742 3.580851    perc
#> 10 0.9 3.205233 3.105796 2.621324 3.536015    perc
#> 11 1.0 3.144158 3.058772 2.588731 3.498018    perc
#> 12 1.1 3.087324 2.990658 2.508082 3.448327    perc
#> 13 1.2 3.034412 2.939828 2.463912 3.413085    perc
#> 14 1.3 2.985118 2.892905 2.409359 3.381694    perc
#> 15 1.4 2.939157 2.849510 2.357536 3.352060    perc
#> 16 1.5 2.896264 2.809304 2.310613 3.324040    perc
#> 17 1.6 2.856191 2.771986 2.267169 3.297502    perc
#> 18 1.7 2.818711 2.737286 2.224618 3.272327    perc
#> 19 1.8 2.783616 2.704966 2.183378 3.248407    perc
#> 20 1.9 2.750715 2.674812 2.145958 3.225646    perc
#> 21 2.0 2.719832 2.646633 2.111833 3.203956    perc
#> 22 2.1 2.690809 2.620258 2.080616 3.182688    perc
#> 23 2.2 2.663500 2.595534 2.052014 3.162345    perc
#> 24 2.3 2.637774 2.572325 2.025763 3.147379    perc
#> 25 2.4 2.613511 2.550506 2.001630 3.134214    perc
#> 26 2.5 2.590601 2.529967 1.979404 3.121707    perc
#> 27 2.6 2.568944 2.510607 1.958902 3.109793    perc
#> 28 2.7 2.548450 2.492336 1.939956 3.098418    perc
#> 29 2.8 2.529037 2.475072 1.922420 3.087533    perc
#> 30 2.9 2.510629 2.458740 1.906160 3.077094    perc
#> 31 3.0 2.493156 2.443273 1.891061 3.067062    perc
#> 
#> $`Diversity profiles`$`Diversity profiles`[[1]]$Medium
#>      q observed     mean    lower    upper ci.type
#> 1  0.0 4.000000 4.000000 4.000000 4.000000    perc
#> 2  0.1 3.902761 3.888094 3.775588 3.955779    perc
#> 3  0.2 3.812649 3.787091 3.588650 3.913503    perc
#> 4  0.3 3.729325 3.695911 3.432986 3.873134    perc
#> 5  0.4 3.652407 3.613518 3.302012 3.834627    perc
#> 6  0.5 3.581486 3.538941 3.191647 3.797927    perc
#> 7  0.6 3.516139 3.471298 3.097955 3.762975    perc
#> 8  0.7 3.455945 3.409800 3.017710 3.729706    perc
#> 9  0.8 3.400487 3.353744 2.948304 3.698053    perc
#> 10 0.9 3.349367 3.302513 2.887658 3.667945    perc
#> 11 1.0 3.302206 3.255565 2.834168 3.639311    perc
#> 12 1.1 3.258648 3.212427 2.786671 3.612079    perc
#> 13 1.2 3.218365 3.172683 2.739885 3.586178    perc
#> 14 1.3 3.181053 3.135971 2.687069 3.561539    perc
#> 15 1.4 3.146434 3.101973 2.638480 3.538091    perc
#> 16 1.5 3.114256 3.070411 2.593566 3.515770    perc
#> 17 1.6 3.084291 3.041041 2.551966 3.494511    perc
#> 18 1.7 3.056333 3.013648 2.513363 3.474252    perc
#> 19 1.8 3.030197 2.988043 2.477479 3.454934    perc
#> 20 1.9 3.005715 2.964059 2.444069 3.436502    perc
#> 21 2.0 2.982739 2.941549 2.416955 3.418902    perc
#> 22 2.1 2.961134 2.920381 2.393052 3.402084    perc
#> 23 2.2 2.940781 2.900440 2.370505 3.386001    perc
#> 24 2.3 2.921573 2.881622 2.349204 3.370608    perc
#> 25 2.4 2.903412 2.863835 2.329053 3.355864    perc
#> 26 2.5 2.886214 2.846996 2.309966 3.341729    perc
#> 27 2.6 2.869900 2.831031 2.291870 3.328166    perc
#> 28 2.7 2.854401 2.815873 2.274696 3.315142    perc
#> 29 2.8 2.839655 2.801463 2.258386 3.302625    perc
#> 30 2.9 2.825604 2.787747 2.242884 3.290584    perc
#> 31 3.0 2.812200 2.774674 2.228141 3.278992    perc
#> 
#> $`Diversity profiles`$`Diversity profiles`[[1]]$Short
#>      q observed     mean    lower upper ci.type
#> 1  0.0 3.000000 2.990000 3.000000     3    perc
#> 2  0.1 2.989707 2.967931 2.889737     3    perc
#> 3  0.2 2.979386 2.946561 2.790512     3    perc
#> 4  0.3 2.969042 2.925888 2.701627     3    perc
#> 5  0.4 2.958682 2.905908 2.622275     3    perc
#> 6  0.5 2.948312 2.886610 2.551599     3    perc
#> 7  0.6 2.937939 2.867981 2.486784     3    perc
#> 8  0.7 2.927570 2.850005 2.425346     3    perc
#> 9  0.8 2.917212 2.832665 2.369193     3    perc
#> 10 0.9 2.906871 2.815943 2.317792     3    perc
#> 11 1.0 2.896553 2.799822 2.270659     3    perc
#> 12 1.1 2.886267 2.784282 2.227362     3    perc
#> 13 1.2 2.876017 2.769305 2.187515     3    perc
#> 14 1.3 2.865812 2.754873 2.150777     3    perc
#> 15 1.4 2.855656 2.740967 2.116848     3    perc
#> 16 1.5 2.845557 2.727569 2.085460     3    perc
#> 17 1.6 2.835521 2.714661 2.056379     3    perc
#> 18 1.7 2.825553 2.702224 2.029395     3    perc
#> 19 1.8 2.815659 2.690241 2.004322     3    perc
#> 20 1.9 2.805845 2.678696 1.980994     3    perc
#> 21 2.0 2.796117 2.667570 1.959264     3    perc
#> 22 2.1 2.786478 2.656847 1.938997     3    perc
#> 23 2.2 2.776935 2.646512 1.920074     3    perc
#> 24 2.3 2.767491 2.636547 1.902387     3    perc
#> 25 2.4 2.758150 2.626940 1.885837     3    perc
#> 26 2.5 2.748918 2.617673 1.870336     3    perc
#> 27 2.6 2.739797 2.608734 1.855802     3    perc
#> 28 2.7 2.730792 2.600108 1.842163     3    perc
#> 29 2.8 2.721904 2.591782 1.829351     3    perc
#> 30 2.9 2.713137 2.583744 1.817304     3    perc
#> 31 3.0 2.704494 2.575982 1.805967     3    perc
#> 
#> attr(,"R")
#> [1] 100
#> attr(,"conf")
#> [1] 0.95
#> attr(,"parameter")
#> [1] "hill"
#> attr(,"ci.type")
#> [1] "perc"
#> 
#> $`Diversity profiles`$`Diversity profiles`[[2]]
#> $`Diversity profiles`$`Diversity profiles`[[2]]$Long
#>      q  observed      mean     lower    upper ci.type
#> 1  0.0 1.3862944 1.3776639 1.0986123 1.386294    perc
#> 2  0.1 1.3574818 1.3433023 1.0863001 1.371606    perc
#> 3  0.2 1.3296293 1.3107059 1.0740771 1.357153    perc
#> 4  0.3 1.3028014 1.2799039 1.0619657 1.342911    perc
#> 5  0.4 1.2770445 1.2508831 1.0499875 1.328911    perc
#> 6  0.5 1.2523869 1.2235985 1.0381628 1.315165    perc
#> 7  0.6 1.2288406 1.1979832 1.0265108 1.301687    perc
#> 8  0.7 1.2064030 1.1739569 1.0150490 1.288487    perc
#> 9  0.8 1.1850594 1.1514313 0.9950056 1.275576    perc
#> 10 0.9 1.1647848 1.1303151 0.9636499 1.262960    perc
#> 11 1.0 1.1455462 1.1149993 0.9511355 1.252136    perc
#> 12 1.1 1.1273048 1.0919483 0.9195184 1.237809    perc
#> 13 1.2 1.1100177 1.0745232 0.9017366 1.227542    perc
#> 14 1.3 1.0936393 1.0581608 0.8793530 1.218315    perc
#> 15 1.4 1.0781229 1.0427847 0.8576155 1.209527    perc
#> 16 1.5 1.0634215 1.0283239 0.8375128 1.201146    perc
#> 17 1.6 1.0494889 1.0147122 0.8185308 1.193142    perc
#> 18 1.7 1.0362798 1.0018881 0.7995841 1.185488    perc
#> 19 1.8 1.0237510 0.9897951 0.7808731 1.178159    perc
#> 20 1.9 1.0118608 0.9783810 0.7635855 1.171132    perc
#> 21 2.0 1.0005702 0.9675977 0.7475541 1.164386    perc
#> 22 2.1 0.9898419 0.9574008 0.7326594 1.157725    perc
#> 23 2.2 0.9796412 0.9477497 0.7188135 1.151307    perc
#> 24 2.3 0.9699355 0.9386070 0.7059340 1.146562    perc
#> 25 2.4 0.9606945 0.9299381 0.6939445 1.142370    perc
#> 26 2.5 0.9518897 0.9217114 0.6827739 1.138371    perc
#> 27 2.6 0.9434949 0.9138976 0.6723569 1.134547    perc
#> 28 2.7 0.9354855 0.9064699 0.6626333 1.130883    perc
#> 29 2.8 0.9278387 0.8994033 0.6535476 1.127363    perc
#> 30 2.9 0.9205332 0.8926751 0.6450493 1.123976    perc
#> 31 3.0 0.9135494 0.8862640 0.6370918 1.120710    perc
#> 
#> $`Diversity profiles`$`Diversity profiles`[[2]]$Medium
#>      q observed     mean     lower    upper ci.type
#> 1  0.0 1.386294 1.386294 1.3862944 1.386294    perc
#> 2  0.1 1.361684 1.357851 1.3285561 1.375177    perc
#> 3  0.2 1.338324 1.331364 1.2777756 1.364433    perc
#> 4  0.3 1.316227 1.306777 1.2334292 1.354064    perc
#> 5  0.4 1.295386 1.283997 1.1945304 1.344072    perc
#> 6  0.5 1.275778 1.262908 1.1605352 1.334455    perc
#> 7  0.6 1.257364 1.243384 1.1307406 1.325210    perc
#> 8  0.7 1.240096 1.225298 1.1044969 1.316329    perc
#> 9  0.8 1.223919 1.208527 1.0812295 1.307806    perc
#> 10 0.9 1.208771 1.192953 1.0604456 1.299631    perc
#> 11 1.0 1.194591 1.178469 1.0417483 1.291793    perc
#> 12 1.1 1.181312 1.164976 1.0248472 1.284282    perc
#> 13 1.2 1.168873 1.152382 1.0079160 1.277086    perc
#> 14 1.3 1.157212 1.140607 0.9884436 1.270191    perc
#> 15 1.4 1.146270 1.129578 0.9701708 1.263585    perc
#> 16 1.5 1.135990 1.119228 0.9529624 1.257255    perc
#> 17 1.6 1.126322 1.109499 0.9367422 1.251190    perc
#> 18 1.7 1.117216 1.100338 0.9214417 1.245374    perc
#> 19 1.8 1.108628 1.091697 0.9069985 1.239798    perc
#> 20 1.9 1.100516 1.083533 0.8933557 1.234447    perc
#> 21 2.0 1.092842 1.075809 0.8821819 1.229312    perc
#> 22 2.1 1.085572 1.068490 0.8722459 1.224379    perc
#> 23 2.2 1.078675 1.061544 0.8627841 1.219639    perc
#> 24 2.3 1.072122 1.054943 0.8537632 1.215081    perc
#> 25 2.4 1.065887 1.048663 0.8451546 1.210695    perc
#> 26 2.5 1.059945 1.042681 0.8369330 1.206473    perc
#> 27 2.6 1.054277 1.036975 0.8290757 1.202404    perc
#> 28 2.7 1.048862 1.031526 0.8215624 1.198481    perc
#> 29 2.8 1.043682 1.026318 0.8143745 1.194695    perc
#> 30 2.9 1.038722 1.021335 0.8074950 1.191040    perc
#> 31 3.0 1.033967 1.016563 0.8009082 1.187509    perc
#> 
#> $`Diversity profiles`$`Diversity profiles`[[2]]$Short
#>      q  observed      mean     lower    upper ci.type
#> 1  0.0 1.0986123 1.0945576 1.0986123 1.098612    perc
#> 2  0.1 1.0951755 1.0871140 1.0610676 1.098612    perc
#> 3  0.2 1.0917172 1.0798005 1.0258577 1.098612    perc
#> 4  0.3 1.0882393 1.0726291 0.9930830 1.098612    perc
#> 5  0.4 1.0847438 1.0656097 0.9627695 1.098612    perc
#> 6  0.5 1.0812328 1.0587495 0.9348800 1.098612    perc
#> 7  0.6 1.0777084 1.0520537 0.9085928 1.098612    perc
#> 8  0.7 1.0741728 1.0455256 0.8831146 1.098612    perc
#> 9  0.8 1.0706283 1.0391672 0.8592836 1.098612    perc
#> 10 0.9 1.0670771 1.0329794 0.8370055 1.098612    perc
#> 11 1.0 1.0635215 1.0269620 0.8161803 1.098612    perc
#> 12 1.1 1.0599638 1.0211142 0.7967077 1.098612    perc
#> 13 1.2 1.0564064 1.0154346 0.7784902 1.098612    perc
#> 14 1.3 1.0528516 1.0099211 0.7614357 1.098612    perc
#> 15 1.4 1.0493016 1.0045715 0.7454577 1.098612    perc
#> 16 1.5 1.0457588 0.9993831 0.7304765 1.098612    perc
#> 17 1.6 1.0422255 0.9943529 0.7164188 1.098612    perc
#> 18 1.7 1.0387040 0.9894777 0.7032173 1.098612    perc
#> 19 1.8 1.0351963 0.9847540 0.6908104 1.098612    perc
#> 20 1.9 1.0317048 0.9801782 0.6791418 1.098612    perc
#> 21 2.0 1.0282315 0.9757466 0.6681599 1.098612    perc
#> 22 2.1 1.0247784 0.9714553 0.6578173 1.098612    perc
#> 23 2.2 1.0213476 0.9673003 0.6480705 1.098612    perc
#> 24 2.3 1.0179410 0.9632777 0.6388793 1.098612    perc
#> 25 2.4 1.0145603 0.9593834 0.6302068 1.098612    perc
#> 26 2.5 1.0112074 0.9556132 0.6220186 1.098612    perc
#> 27 2.6 1.0078840 0.9519633 0.6142832 1.098612    perc
#> 28 2.7 1.0045915 0.9484296 0.6069710 1.098612    perc
#> 29 2.8 1.0013316 0.9450081 0.6000549 1.098612    perc
#> 30 2.9 0.9981055 0.9416949 0.5935094 1.098612    perc
#> 31 3.0 0.9949147 0.9384863 0.5873111 1.098612    perc
#> 
#> attr(,"R")
#> [1] 100
#> attr(,"conf")
#> [1] 0.95
#> attr(,"parameter")
#> [1] "renyi"
#> attr(,"ci.type")
#> [1] "perc"
#> 
#> $`Diversity profiles`$`Diversity profiles`[[3]]
#> $`Diversity profiles`$`Diversity profiles`[[3]]$Long
#>      q  observed      mean     lower     upper ci.type
#> 1  0.0 3.0000000 2.9700000 2.0000000 3.0000000    perc
#> 2  0.1 2.6589612 2.6142899 1.8425073 2.7071922    perc
#> 3  0.2 2.3713503 2.3194816 1.7017745 2.4519732    perc
#> 4  0.3 2.1274272 2.0730801 1.5757541 2.2286838    perc
#> 5  0.4 1.9193870 1.8654543 1.4626693 2.0327394    perc
#> 6  0.5 1.7409539 1.6891293 1.3609760 1.8602439    perc
#> 7  0.6 1.5870644 1.5382632 1.2693311 1.7079121    perc
#> 8  0.7 1.4536179 1.4082595 1.1865643 1.5729674    perc
#> 9  0.8 1.3372809 1.2954757 1.1009168 1.4530572    perc
#> 10 0.9 1.2353333 1.1970062 1.0116123 1.3461839    perc
#> 11 1.0 1.1455462 1.1149993 0.9511355 1.2521362    perc
#> 12 1.1 1.0660859 1.0341214 0.8785092 1.1642592    perc
#> 13 1.2 0.9954373 0.9662830 0.8250970 1.0884547    perc
#> 14 1.3 0.9323437 0.9057441 0.7729228 1.0204759    perc
#> 15 1.4 0.8757575 0.8514676 0.7259860 0.9589139    perc
#> 16 1.5 0.8248023 0.8025936 0.6842711 0.9029960    perc
#> 17 1.6 0.7787414 0.7584049 0.6467636 0.8520611    perc
#> 18 1.7 0.7369530 0.7182997 0.6123204 0.8055410    perc
#> 19 1.8 0.6989100 0.6817704 0.5807215 0.7629456    perc
#> 20 1.9 0.6641637 0.6483867 0.5522566 0.7238502    perc
#> 21 2.0 0.6323302 0.6177816 0.5264757 0.6878858    perc
#> 22 2.1 0.6030805 0.5896411 0.5030209 0.6546802    perc
#> 23 2.2 0.5761307 0.5636949 0.4816013 0.6240113    perc
#> 24 2.3 0.5512352 0.5397094 0.4619696 0.5959568    perc
#> 25 2.4 0.5281808 0.5174818 0.4439149 0.5699732    perc
#> 26 2.5 0.5067815 0.4968355 0.4272563 0.5457922    perc
#> 27 2.6 0.4868749 0.4776160 0.4118385 0.5232521    perc
#> 28 2.7 0.4683181 0.4596876 0.3975273 0.5022090    perc
#> 29 2.8 0.4509853 0.4429309 0.3842065 0.4825347    perc
#> 30 2.9 0.4347655 0.4272400 0.3717750 0.4641141    perc
#> 31 3.0 0.4195602 0.4125215 0.3601446 0.4468443    perc
#> 
#> $`Diversity profiles`$`Diversity profiles`[[3]]$Medium
#>      q  observed      mean     lower     upper ci.type
#> 1  0.0 3.0000000 3.0000000 3.0000000 3.0000000    perc
#> 2  0.1 2.6732473 2.6604222 2.5620810 2.7194843    perc
#> 3  0.2 2.3966283 2.3769233 2.2242003 2.4735957    perc
#> 4  0.3 2.1610045 2.1381256 1.9588736 2.2573471    perc
#> 5  0.4 1.9590700 1.9352680 1.7461723 2.0665450    perc
#> 6  0.5 1.7849628 1.7615529 1.5730345 1.8976542    perc
#> 7  0.6 1.6339616 1.6116660 1.4297949 1.7476872    perc
#> 8  0.7 1.5022489 1.4814204 1.3094869 1.6141132    perc
#> 9  0.8 1.3867251 1.3674921 1.2070382 1.4947823    perc
#> 10 0.9 1.2848625 1.2672226 1.1187143 1.3878635    perc
#> 11 1.0 1.1945906 1.1784694 1.0417483 1.2917934    perc
#> 12 1.1 1.1142058 1.0994932 0.9740806 1.2052331    perc
#> 13 1.2 1.0422993 1.0288720 0.9128221 1.1270330    perc
#> 14 1.3 0.9777011 0.9654350 0.8553616 1.0562035    perc
#> 15 1.4 0.9194342 0.9082117 0.8040763 0.9918900    perc
#> 16 1.5 0.8666793 0.8563926 0.7580486 0.9333528    perc
#> 17 1.6 0.8187454 0.8092977 0.7165551 0.8799497    perc
#> 18 1.7 0.7750476 0.7663530 0.6789942 0.8311216    perc
#> 19 1.8 0.7350876 0.7270707 0.6448624 0.7863805    perc
#> 20 1.9 0.6984396 0.6910342 0.6137352 0.7452995    perc
#> 21 2.0 0.6647377 0.6578858 0.5859857 0.7075039    perc
#> 22 2.1 0.6336659 0.6273167 0.5606855 0.6726640    perc
#> 23 2.2 0.6049505 0.5990594 0.5372760 0.6404892    perc
#> 24 2.3 0.5783532 0.5728809 0.5155619 0.6107226    perc
#> 25 2.4 0.5536656 0.5485772 0.4953738 0.5831364    perc
#> 26 2.5 0.5307051 0.5259695 0.4765636 0.5575283    perc
#> 27 2.6 0.5093107 0.5049002 0.4590013 0.5337183    perc
#> 28 2.7 0.4893400 0.4852297 0.4425726 0.5115456    perc
#> 29 2.8 0.4706669 0.4668343 0.4271761 0.4908667    perc
#> 30 2.9 0.4531792 0.4496039 0.4127220 0.4715529    perc
#> 31 3.0 0.4367766 0.4334401 0.3991301 0.4534885    perc
#> 
#> $`Diversity profiles`$`Diversity profiles`[[3]]$Short
#>      q  observed      mean     lower     upper ci.type
#> 1  0.0 2.0000000 1.9900000 2.0000000 2.0000000    perc
#> 2  0.1 1.8661938 1.8464914 1.7764163 1.8754171    perc
#> 3  0.2 1.7437218 1.7169440 1.5907408 1.7602809    perc
#> 4  0.3 1.6315129 1.5997386 1.4354050 1.6538133    perc
#> 5  0.4 1.5286044 1.4934713 1.3044741 1.5553034    perc
#> 6  0.5 1.4341299 1.3969210 1.1932777 1.4641016    perc
#> 7  0.6 1.3473096 1.3090229 1.0970435 1.3796139    perc
#> 8  0.7 1.2674411 1.2288460 1.0122727 1.3012972    perc
#> 9  0.8 1.1938915 1.1555741 0.9383166 1.2286547    perc
#> 10 0.9 1.1260900 1.0884900 0.8734265 1.1612317    perc
#> 11 1.0 1.0635215 1.0269620 0.8161803 1.0986123    perc
#> 12 1.1 1.0057210 0.9704325 0.7654155 1.0404154    perc
#> 13 1.2 0.9522684 0.9184080 0.7201765 0.9862922    perc
#> 14 1.3 0.9027839 0.8704514 0.6796737 0.9359230    perc
#> 15 1.4 0.8569240 0.8261741 0.6432513 0.8890150    perc
#> 16 1.5 0.8143777 0.7852303 0.6103613 0.8452995    perc
#> 17 1.6 0.7748634 0.7473119 0.5805439 0.8045302    perc
#> 18 1.7 0.7381257 0.7121434 0.5534114 0.7664813    perc
#> 19 1.8 0.7039330 0.6794785 0.5286348 0.7309454    perc
#> 20 1.9 0.6720751 0.6490963 0.5059340 0.6977322    perc
#> 21 2.0 0.6423611 0.6207986 0.4850694 0.6666667    perc
#> 22 2.1 0.6146175 0.5944071 0.4658352 0.6375883    perc
#> 23 2.2 0.5886863 0.5697612 0.4480536 0.6103496    perc
#> 24 2.3 0.5644238 0.5467162 0.4315706 0.5848146    perc
#> 25 2.4 0.5416992 0.5251411 0.4162523 0.5608586    perc
#> 26 2.5 0.5203930 0.5049177 0.4019819 0.5383666    perc
#> 27 2.6 0.5003964 0.4859388 0.3886570 0.5172329    perc
#> 28 2.7 0.4816099 0.4681071 0.3761876 0.4973602    perc
#> 29 2.8 0.4639427 0.4513343 0.3644944 0.4786586    perc
#> 30 2.9 0.4473116 0.4355400 0.3535073 0.4610454    perc
#> 31 3.0 0.4316406 0.4206510 0.3431641 0.4444444    perc
#> 
#> attr(,"R")
#> [1] 100
#> attr(,"conf")
#> [1] 0.95
#> attr(,"parameter")
#> [1] "tsallis"
#> attr(,"ci.type")
#> [1] "perc"
#> 
#> 
#> $`Diversity profiles`$Warnings
#> $`Diversity profiles`$Warnings$hill
#> NULL
#> 
#> $`Diversity profiles`$Warnings$renyi
#> NULL
#> 
#> $`Diversity profiles`$Warnings$tsallis
#> NULL
#> 
#> 
#>