Perform Pairwise t Tests of Adjusted Means from augmentedRCBD Output

pairwise.augmentedRCBD performs pairwise t tests of adjusted means from an object of class augmentedRCBD.

Usage

pairwise.augmentedRCBD(
  aug,
  cl = NULL,
  alpha = NULL,
  p.adjust = c("none", "tukey", "sidak")
)

Arguments

aug

An object of class augmentedRCBD.

cl

A cluster object created by makeCluster for parallel evaluations.

alpha

Type I error probability (Significance level).

p.adjust

The p value adjustment method. Either "none", "tukey" or "sidak".

Value

A data frame of pairwise comparisons of treatments.

Details

The default pairwise comparison in augmentedRCBD employs pairs.emmGrid function from emmeans which is very slow for large number of comparisons. This function attempts to do the same faster with parallel computing with the package parallel-package.

See also

Examples

library(augmentedRCBD)
#> 
#> --------------------------------------------------------------------------------
#> Welcome to augmentedRCBD version 0.1.7
#> 
#> 
#> # To know how to use this package type:
#>   browseVignettes(package = 'augmentedRCBD')
#>   for the package vignette.
#> 
#> # To know whats new in this version type:
#>   news(package='augmentedRCBD')
#>   for the NEWS file.
#> 
#> # To cite the methods in the package type:
#>   citation(package='augmentedRCBD')
#> 
#> # To suppress this message use:
#>   suppressPackageStartupMessages(library(augmentedRCBD))
#> --------------------------------------------------------------------------------

blk <- c(rep(1,7),rep(2,6),rep(3,7))
trt <- c(1, 2, 3, 4, 7, 11, 12, 1, 2, 3, 4, 5, 9, 1, 2, 3, 4, 8, 6, 10)
y1 <- c(92, 79, 87, 81, 96, 89, 82, 79, 81, 81, 91, 79, 78, 83, 77, 78, 78,
        70, 75, 74)
y2 <- c(258, 224, 238, 278, 347, 300, 289, 260, 220, 237, 227, 281, 311, 250,
        240, 268, 287, 226, 395, 450)
data <- data.frame(blk, trt, y1, y2)
# Convert block and treatment to factors
data$blk <- as.factor(data$blk)
data$trt <- as.factor(data$trt)
# Results for variable y1
out1 <- augmentedRCBD(data$blk, data$trt, data$y1, method.comp = "lsd",
                      alpha = 0.05, group = TRUE, console = TRUE)
#> 
#> Augmented Design Details
#> ========================
#>                                        
#> Number of blocks           "3"         
#> Number of treatments       "12"        
#> Number of check treatments "4"         
#> Number of test treatments  "8"         
#> Check treatments           "1, 2, 3, 4"
#> 
#> 
#> ANOVA, Treatment Adjusted
#> =========================
#>                                      Df Sum Sq Mean Sq F value Pr(>F)  
#> Block (ignoring Treatments)           2  360.1  180.04   6.675 0.0298 *
#> Treatment (eliminating Blocks)       11  285.1   25.92   0.961 0.5499  
#>   Treatment: Check                    3   52.9   17.64   0.654 0.6092  
#>   Treatment: Test and Test vs. Check  8  232.2   29.02   1.076 0.4779  
#> Residuals                             6  161.8   26.97                 
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 
#> ANOVA, Block Adjusted
#> =====================
#>                                Df Sum Sq Mean Sq F value Pr(>F)
#> Treatment (ignoring Blocks)    11  575.7   52.33   1.940  0.215
#>   Treatment: Check              3   52.9   17.64   0.654  0.609
#>   Treatment: Test               7  505.9   72.27   2.679  0.125
#>   Treatment: Test vs. Check     1   16.9   16.87   0.626  0.459
#> Block (eliminating Treatments)  2   69.5   34.75   1.288  0.342
#> Residuals                       6  161.8   26.97               
#> 
#> Coefficient of Variation
#> ========================
#> 6.372367
#> 
#> Overall Adjusted Mean
#> =====================
#> 81.0625
#> 
#> Standard Errors
#> ===============
#>                                          Std. Error of Diff.  CD (5%)
#> Control Treatment Means                             4.240458 10.37603
#> Two Test Treatments (Same Block)                    7.344688 17.97180
#> Two Test Treatments (Different Blocks)              8.211611 20.09309
#> A Test Treatment and a Control Treatment            6.704752 16.40594
#> 
#> Treatment Means
#> ===============
#>  Treatment Block Means   SE r   Min   Max Adjusted Means
#>          1       84.67 3.84 3 79.00 92.00          84.67
#>         10     3 74.00 <NA> 1 74.00 74.00          77.25
#>         11     1 89.00 <NA> 1 89.00 89.00          86.50
#>         12     1 82.00 <NA> 1 82.00 82.00          79.50
#>          2       79.00 1.15 3 77.00 81.00          79.00
#>          3       82.00 2.65 3 78.00 87.00          82.00
#>          4       83.33 3.93 3 78.00 91.00          83.33
#>          5     2 79.00 <NA> 1 79.00 79.00          78.25
#>          6     3 75.00 <NA> 1 75.00 75.00          78.25
#>          7     1 96.00 <NA> 1 96.00 96.00          93.50
#>          8     3 70.00 <NA> 1 70.00 70.00          73.25
#>          9     2 78.00 <NA> 1 78.00 78.00          77.25
#> 
#> 
#> Comparisons
#> ===========
#> 
#> Method : lsd
#> 
#>                   contrast estimate   SE df t.ratio p.value sig
#>    treatment1 - treatment2     5.67 4.24  6   1.336   0.230    
#>    treatment1 - treatment3     2.67 4.24  6   0.629   0.553    
#>    treatment1 - treatment4     1.33 4.24  6   0.314   0.764    
#>    treatment1 - treatment5     6.42 6.36  6   1.009   0.352    
#>    treatment1 - treatment6     6.42 6.36  6   1.009   0.352    
#>    treatment1 - treatment7    -8.83 6.36  6  -1.389   0.214    
#>    treatment1 - treatment8    11.42 6.36  6   1.795   0.123    
#>    treatment1 - treatment9     7.42 6.36  6   1.166   0.288    
#>   treatment1 - treatment10     7.42 6.36  6   1.166   0.288    
#>   treatment1 - treatment11    -1.83 6.36  6  -0.288   0.783    
#>   treatment1 - treatment12     5.17 6.36  6   0.812   0.448    
#>    treatment2 - treatment3    -3.00 4.24  6  -0.707   0.506    
#>    treatment2 - treatment4    -4.33 4.24  6  -1.022   0.346    
#>    treatment2 - treatment5     0.75 6.36  6   0.118   0.910    
#>    treatment2 - treatment6     0.75 6.36  6   0.118   0.910    
#>    treatment2 - treatment7   -14.50 6.36  6  -2.280   0.063    
#>    treatment2 - treatment8     5.75 6.36  6   0.904   0.401    
#>    treatment2 - treatment9     1.75 6.36  6   0.275   0.792    
#>   treatment2 - treatment10     1.75 6.36  6   0.275   0.792    
#>   treatment2 - treatment11    -7.50 6.36  6  -1.179   0.283    
#>   treatment2 - treatment12    -0.50 6.36  6  -0.079   0.940    
#>    treatment3 - treatment4    -1.33 4.24  6  -0.314   0.764    
#>    treatment3 - treatment5     3.75 6.36  6   0.590   0.577    
#>    treatment3 - treatment6     3.75 6.36  6   0.590   0.577    
#>    treatment3 - treatment7   -11.50 6.36  6  -1.808   0.121    
#>    treatment3 - treatment8     8.75 6.36  6   1.376   0.218    
#>    treatment3 - treatment9     4.75 6.36  6   0.747   0.483    
#>   treatment3 - treatment10     4.75 6.36  6   0.747   0.483    
#>   treatment3 - treatment11    -4.50 6.36  6  -0.707   0.506    
#>   treatment3 - treatment12     2.50 6.36  6   0.393   0.708    
#>    treatment4 - treatment5     5.08 6.36  6   0.799   0.455    
#>    treatment4 - treatment6     5.08 6.36  6   0.799   0.455    
#>    treatment4 - treatment7   -10.17 6.36  6  -1.598   0.161    
#>    treatment4 - treatment8    10.08 6.36  6   1.585   0.164    
#>    treatment4 - treatment9     6.08 6.36  6   0.956   0.376    
#>   treatment4 - treatment10     6.08 6.36  6   0.956   0.376    
#>   treatment4 - treatment11    -3.17 6.36  6  -0.498   0.636    
#>   treatment4 - treatment12     3.83 6.36  6   0.603   0.569    
#>    treatment5 - treatment6     0.00 8.21  6   0.000   1.000    
#>    treatment5 - treatment7   -15.25 8.21  6  -1.857   0.113    
#>    treatment5 - treatment8     5.00 8.21  6   0.609   0.565    
#>    treatment5 - treatment9     1.00 7.34  6   0.136   0.896    
#>   treatment5 - treatment10     1.00 8.21  6   0.122   0.907    
#>   treatment5 - treatment11    -8.25 8.21  6  -1.005   0.354    
#>   treatment5 - treatment12    -1.25 8.21  6  -0.152   0.884    
#>    treatment6 - treatment7   -15.25 8.21  6  -1.857   0.113    
#>    treatment6 - treatment8     5.00 7.34  6   0.681   0.521    
#>    treatment6 - treatment9     1.00 8.21  6   0.122   0.907    
#>   treatment6 - treatment10     1.00 7.34  6   0.136   0.896    
#>   treatment6 - treatment11    -8.25 8.21  6  -1.005   0.354    
#>   treatment6 - treatment12    -1.25 8.21  6  -0.152   0.884    
#>    treatment7 - treatment8    20.25 8.21  6   2.466   0.049   *
#>    treatment7 - treatment9    16.25 8.21  6   1.979   0.095    
#>   treatment7 - treatment10    16.25 8.21  6   1.979   0.095    
#>   treatment7 - treatment11     7.00 7.34  6   0.953   0.377    
#>   treatment7 - treatment12    14.00 7.34  6   1.906   0.105    
#>    treatment8 - treatment9    -4.00 8.21  6  -0.487   0.643    
#>   treatment8 - treatment10    -4.00 7.34  6  -0.545   0.606    
#>   treatment8 - treatment11   -13.25 8.21  6  -1.614   0.158    
#>   treatment8 - treatment12    -6.25 8.21  6  -0.761   0.475    
#>   treatment9 - treatment10     0.00 8.21  6   0.000   1.000    
#>   treatment9 - treatment11    -9.25 8.21  6  -1.126   0.303    
#>   treatment9 - treatment12    -2.25 8.21  6  -0.274   0.793    
#>  treatment10 - treatment11    -9.25 8.21  6  -1.126   0.303    
#>  treatment10 - treatment12    -2.25 8.21  6  -0.274   0.793    
#>  treatment11 - treatment12     7.00 7.34  6   0.953   0.377    
#> 
#> Treatment Groups
#> ================
#> 
#> Method : lsd
#> 
#>  Treatment Adjusted Means   SE df lower.CL upper.CL Group
#>          8          73.25 5.61  6    59.52    86.98    1 
#>          9          77.25 5.61  6    63.52    90.98    12
#>         10          77.25 5.61  6    63.52    90.98    12
#>          5          78.25 5.61  6    64.52    91.98    12
#>          6          78.25 5.61  6    64.52    91.98    12
#>          2          79.00 3.00  6    71.66    86.34    12
#>         12          79.50 5.61  6    65.77    93.23    12
#>          3          82.00 3.00  6    74.66    89.34    12
#>          4          83.33 3.00  6    76.00    90.67    12
#>          1          84.67 3.00  6    77.33    92.00    12
#>         11          86.50 5.61  6    72.77   100.23    12
#>          7          93.50 5.61  6    79.77   107.23     2
# Results for variable y2
out2 <- augmentedRCBD(data$blk, data$trt, data$y1, method.comp = "lsd",
                      alpha = 0.05, group = TRUE, console = TRUE)
#> 
#> Augmented Design Details
#> ========================
#>                                        
#> Number of blocks           "3"         
#> Number of treatments       "12"        
#> Number of check treatments "4"         
#> Number of test treatments  "8"         
#> Check treatments           "1, 2, 3, 4"
#> 
#> 
#> ANOVA, Treatment Adjusted
#> =========================
#>                                      Df Sum Sq Mean Sq F value Pr(>F)  
#> Block (ignoring Treatments)           2  360.1  180.04   6.675 0.0298 *
#> Treatment (eliminating Blocks)       11  285.1   25.92   0.961 0.5499  
#>   Treatment: Check                    3   52.9   17.64   0.654 0.6092  
#>   Treatment: Test and Test vs. Check  8  232.2   29.02   1.076 0.4779  
#> Residuals                             6  161.8   26.97                 
#> ---
#> Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1
#> 
#> ANOVA, Block Adjusted
#> =====================
#>                                Df Sum Sq Mean Sq F value Pr(>F)
#> Treatment (ignoring Blocks)    11  575.7   52.33   1.940  0.215
#>   Treatment: Check              3   52.9   17.64   0.654  0.609
#>   Treatment: Test               7  505.9   72.27   2.679  0.125
#>   Treatment: Test vs. Check     1   16.9   16.87   0.626  0.459
#> Block (eliminating Treatments)  2   69.5   34.75   1.288  0.342
#> Residuals                       6  161.8   26.97               
#> 
#> Coefficient of Variation
#> ========================
#> 6.372367
#> 
#> Overall Adjusted Mean
#> =====================
#> 81.0625
#> 
#> Standard Errors
#> ===============
#>                                          Std. Error of Diff.  CD (5%)
#> Control Treatment Means                             4.240458 10.37603
#> Two Test Treatments (Same Block)                    7.344688 17.97180
#> Two Test Treatments (Different Blocks)              8.211611 20.09309
#> A Test Treatment and a Control Treatment            6.704752 16.40594
#> 
#> Treatment Means
#> ===============
#>  Treatment Block Means   SE r   Min   Max Adjusted Means
#>          1       84.67 3.84 3 79.00 92.00          84.67
#>         10     3 74.00 <NA> 1 74.00 74.00          77.25
#>         11     1 89.00 <NA> 1 89.00 89.00          86.50
#>         12     1 82.00 <NA> 1 82.00 82.00          79.50
#>          2       79.00 1.15 3 77.00 81.00          79.00
#>          3       82.00 2.65 3 78.00 87.00          82.00
#>          4       83.33 3.93 3 78.00 91.00          83.33
#>          5     2 79.00 <NA> 1 79.00 79.00          78.25
#>          6     3 75.00 <NA> 1 75.00 75.00          78.25
#>          7     1 96.00 <NA> 1 96.00 96.00          93.50
#>          8     3 70.00 <NA> 1 70.00 70.00          73.25
#>          9     2 78.00 <NA> 1 78.00 78.00          77.25
#> 
#> 
#> Comparisons
#> ===========
#> 
#> Method : lsd
#> 
#>                   contrast estimate   SE df t.ratio p.value sig
#>    treatment1 - treatment2     5.67 4.24  6   1.336   0.230    
#>    treatment1 - treatment3     2.67 4.24  6   0.629   0.553    
#>    treatment1 - treatment4     1.33 4.24  6   0.314   0.764    
#>    treatment1 - treatment5     6.42 6.36  6   1.009   0.352    
#>    treatment1 - treatment6     6.42 6.36  6   1.009   0.352    
#>    treatment1 - treatment7    -8.83 6.36  6  -1.389   0.214    
#>    treatment1 - treatment8    11.42 6.36  6   1.795   0.123    
#>    treatment1 - treatment9     7.42 6.36  6   1.166   0.288    
#>   treatment1 - treatment10     7.42 6.36  6   1.166   0.288    
#>   treatment1 - treatment11    -1.83 6.36  6  -0.288   0.783    
#>   treatment1 - treatment12     5.17 6.36  6   0.812   0.448    
#>    treatment2 - treatment3    -3.00 4.24  6  -0.707   0.506    
#>    treatment2 - treatment4    -4.33 4.24  6  -1.022   0.346    
#>    treatment2 - treatment5     0.75 6.36  6   0.118   0.910    
#>    treatment2 - treatment6     0.75 6.36  6   0.118   0.910    
#>    treatment2 - treatment7   -14.50 6.36  6  -2.280   0.063    
#>    treatment2 - treatment8     5.75 6.36  6   0.904   0.401    
#>    treatment2 - treatment9     1.75 6.36  6   0.275   0.792    
#>   treatment2 - treatment10     1.75 6.36  6   0.275   0.792    
#>   treatment2 - treatment11    -7.50 6.36  6  -1.179   0.283    
#>   treatment2 - treatment12    -0.50 6.36  6  -0.079   0.940    
#>    treatment3 - treatment4    -1.33 4.24  6  -0.314   0.764    
#>    treatment3 - treatment5     3.75 6.36  6   0.590   0.577    
#>    treatment3 - treatment6     3.75 6.36  6   0.590   0.577    
#>    treatment3 - treatment7   -11.50 6.36  6  -1.808   0.121    
#>    treatment3 - treatment8     8.75 6.36  6   1.376   0.218    
#>    treatment3 - treatment9     4.75 6.36  6   0.747   0.483    
#>   treatment3 - treatment10     4.75 6.36  6   0.747   0.483    
#>   treatment3 - treatment11    -4.50 6.36  6  -0.707   0.506    
#>   treatment3 - treatment12     2.50 6.36  6   0.393   0.708    
#>    treatment4 - treatment5     5.08 6.36  6   0.799   0.455    
#>    treatment4 - treatment6     5.08 6.36  6   0.799   0.455    
#>    treatment4 - treatment7   -10.17 6.36  6  -1.598   0.161    
#>    treatment4 - treatment8    10.08 6.36  6   1.585   0.164    
#>    treatment4 - treatment9     6.08 6.36  6   0.956   0.376    
#>   treatment4 - treatment10     6.08 6.36  6   0.956   0.376    
#>   treatment4 - treatment11    -3.17 6.36  6  -0.498   0.636    
#>   treatment4 - treatment12     3.83 6.36  6   0.603   0.569    
#>    treatment5 - treatment6     0.00 8.21  6   0.000   1.000    
#>    treatment5 - treatment7   -15.25 8.21  6  -1.857   0.113    
#>    treatment5 - treatment8     5.00 8.21  6   0.609   0.565    
#>    treatment5 - treatment9     1.00 7.34  6   0.136   0.896    
#>   treatment5 - treatment10     1.00 8.21  6   0.122   0.907    
#>   treatment5 - treatment11    -8.25 8.21  6  -1.005   0.354    
#>   treatment5 - treatment12    -1.25 8.21  6  -0.152   0.884    
#>    treatment6 - treatment7   -15.25 8.21  6  -1.857   0.113    
#>    treatment6 - treatment8     5.00 7.34  6   0.681   0.521    
#>    treatment6 - treatment9     1.00 8.21  6   0.122   0.907    
#>   treatment6 - treatment10     1.00 7.34  6   0.136   0.896    
#>   treatment6 - treatment11    -8.25 8.21  6  -1.005   0.354    
#>   treatment6 - treatment12    -1.25 8.21  6  -0.152   0.884    
#>    treatment7 - treatment8    20.25 8.21  6   2.466   0.049   *
#>    treatment7 - treatment9    16.25 8.21  6   1.979   0.095    
#>   treatment7 - treatment10    16.25 8.21  6   1.979   0.095    
#>   treatment7 - treatment11     7.00 7.34  6   0.953   0.377    
#>   treatment7 - treatment12    14.00 7.34  6   1.906   0.105    
#>    treatment8 - treatment9    -4.00 8.21  6  -0.487   0.643    
#>   treatment8 - treatment10    -4.00 7.34  6  -0.545   0.606    
#>   treatment8 - treatment11   -13.25 8.21  6  -1.614   0.158    
#>   treatment8 - treatment12    -6.25 8.21  6  -0.761   0.475    
#>   treatment9 - treatment10     0.00 8.21  6   0.000   1.000    
#>   treatment9 - treatment11    -9.25 8.21  6  -1.126   0.303    
#>   treatment9 - treatment12    -2.25 8.21  6  -0.274   0.793    
#>  treatment10 - treatment11    -9.25 8.21  6  -1.126   0.303    
#>  treatment10 - treatment12    -2.25 8.21  6  -0.274   0.793    
#>  treatment11 - treatment12     7.00 7.34  6   0.953   0.377    
#> 
#> Treatment Groups
#> ================
#> 
#> Method : lsd
#> 
#>  Treatment Adjusted Means   SE df lower.CL upper.CL Group
#>          8          73.25 5.61  6    59.52    86.98    1 
#>          9          77.25 5.61  6    63.52    90.98    12
#>         10          77.25 5.61  6    63.52    90.98    12
#>          5          78.25 5.61  6    64.52    91.98    12
#>          6          78.25 5.61  6    64.52    91.98    12
#>          2          79.00 3.00  6    71.66    86.34    12
#>         12          79.50 5.61  6    65.77    93.23    12
#>          3          82.00 3.00  6    74.66    89.34    12
#>          4          83.33 3.00  6    76.00    90.67    12
#>          1          84.67 3.00  6    77.33    92.00    12
#>         11          86.50 5.61  6    72.77   100.23    12
#>          7          93.50 5.61  6    79.77   107.23     2

# Make cluster ----
library(parallel)
# Check if running under R CMD check and adjust cores accordingly
if (nzchar(Sys.getenv("_R_CHECK_LIMIT_CORES_"))) {
  ncores <- 2
} else {
  ncores <- max(2L, parallel::detectCores() - 4)
}

cl <- makeCluster(getOption("cl.cores", ncores))

# Pairwise t test without p value adjustment ----
pout1 <- pairwise.augmentedRCBD(out1, cl = cl,
                                p.adjust = "none")
pout1
#>       contrast   estimate       SE df     t.ratio    p.value sig
#> 1     1 - (10)   7.416667 6.704752  6  1.10618056 1.68898284    
#> 2     1 - (11)  -1.833333 6.704752  6 -0.27343789 0.79368531    
#> 3     1 - (12)   5.166667 6.704752  6  0.77059769 1.52980915    
#> 4        1 - 2   5.666667 4.240458  6  1.33633373 1.77011499    
#> 5        1 - 3   2.666667 4.240458  6  0.62886293 1.44738825    
#> 6        1 - 4   1.333333 4.240458  6  0.31443147 1.23616021    
#> 7      1 - (5)   6.416667 6.704752  6  0.95703262 1.62449310    
#> 8      1 - (6)   6.416667 6.704752  6  0.95703262 1.62449310    
#> 9      1 - (7)  -8.833333 6.704752  6 -1.31747347 0.23575127    
#> 10     1 - (8)  11.416667 6.704752  6  1.70277232 1.86049532    
#> 11     1 - (9)   7.416667 6.704752  6  1.10618056 1.68898284    
#> 12 (10) - (11)  -9.250000 8.211611  6 -1.12645375 0.30300033    
#> 13 (10) - (12)  -2.250000 8.211611  6 -0.27400226 0.79327167    
#> 14    (10) - 2  -1.750000 6.704752  6 -0.26100890 0.80281330    
#> 15    (10) - 3  -4.750000 6.704752  6 -0.70845272 0.50524328    
#> 16    (10) - 4  -6.083333 6.704752  6 -0.90731664 0.39921196    
#> 17  (10) - (5)  -1.000000 8.211611  6 -0.12177878 0.90705045    
#> 18  (10) - (6)  -1.000000 7.344688  6 -0.13615282 0.89615381    
#> 19  (10) - (7) -16.250000 8.211611  6 -1.97890523 0.09516823    
#> 20  (10) - (8)   4.000000 7.344688  6  0.54461128 1.39434533    
#> 21  (10) - (9)   0.000000 8.211611  6  0.00000000 1.00000000    
#> 22 (11) - (12)   7.000000 7.344688  6  0.95306973 1.62264457    
#> 23    (11) - 2   7.500000 6.704752  6  1.11860955 1.69391854    
#> 24    (11) - 3   4.500000 6.704752  6  0.67116573 1.47290592    
#> 25    (11) - 4   3.166667 6.704752  6  0.47230181 1.34659999    
#> 26  (11) - (5)   8.250000 8.211611  6  1.00467496 1.64616381    
#> 27  (11) - (6)   8.250000 8.211611  6  1.00467496 1.64616381    
#> 28  (11) - (7)  -7.000000 7.344688  6 -0.95306973 0.37735543    
#> 29  (11) - (8)  13.250000 8.211611  6  1.61356888 1.84225214    
#> 30  (11) - (9)   9.250000 8.211611  6  1.12645375 1.69699967    
#> 31    (12) - 2   0.500000 6.704752  6  0.07457397 1.05702215    
#> 32    (12) - 3  -2.500000 6.704752  6 -0.37286985 0.72206273    
#> 33    (12) - 4  -3.833333 6.704752  6 -0.57173377 0.58826470    
#> 34  (12) - (5)   1.250000 8.211611  6  0.15222348 1.11599954    
#> 35  (12) - (6)   1.250000 8.211611  6  0.15222348 1.11599954    
#> 36  (12) - (7) -14.000000 7.344688  6 -1.90613947 0.10526990    
#> 37  (12) - (8)   6.250000 8.211611  6  0.76111740 1.52457343    
#> 38  (12) - (9)   2.250000 8.211611  6  0.27400226 1.20672833    
#> 39       2 - 3  -3.000000 4.240458  6 -0.70747080 0.50581089    
#> 40       2 - 4  -4.333333 4.240458  6 -1.02190227 0.34624985    
#> 41     2 - (5)   0.750000 6.704752  6  0.11186096 1.08541796    
#> 42     2 - (6)   0.750000 6.704752  6  0.11186096 1.08541796    
#> 43     2 - (7) -14.500000 6.704752  6 -2.16264514 0.07380701    
#> 44     2 - (8)   5.750000 6.704752  6  0.85760066 1.57596055    
#> 45     2 - (9)   1.750000 6.704752  6  0.26100890 1.19718670    
#> 46       3 - 4  -1.333333 4.240458  6 -0.31443147 0.76383979    
#> 47     3 - (5)   3.750000 6.704752  6  0.55930478 1.40380253    
#> 48     3 - (6)   3.750000 6.704752  6  0.55930478 1.40380253    
#> 49     3 - (7) -11.500000 6.704752  6 -1.71520132 0.13713052    
#> 50     3 - (8)   8.750000 6.704752  6  1.30504448 1.76031005    
#> 51     3 - (9)   4.750000 6.704752  6  0.70845272 1.49475672    
#> 52     4 - (5)   5.083333 6.704752  6  0.75816870 1.52293671    
#> 53     4 - (6)   5.083333 6.704752  6  0.75816870 1.52293671    
#> 54     4 - (7) -10.166667 6.704752  6 -1.51633739 0.18022066    
#> 55     4 - (8)  10.083333 6.704752  6  1.50390840 1.81669827    
#> 56     4 - (9)   6.083333 6.704752  6  0.90731664 1.60078804    
#> 57   (5) - (6)   0.000000 8.211611  6  0.00000000 1.00000000    
#> 58   (5) - (7) -15.250000 8.211611  6 -1.85712645 0.11267166    
#> 59   (5) - (8)   5.000000 8.211611  6  0.60889392 1.43507937    
#> 60   (5) - (9)   1.000000 7.344688  6  0.13615282 1.10384619    
#> 61   (6) - (7) -15.250000 8.211611  6 -1.85712645 0.11267166    
#> 62   (6) - (8)   5.000000 7.344688  6  0.68076409 1.47858869    
#> 63   (6) - (9)   1.000000 8.211611  6  0.12177878 1.09294955    
#> 64   (7) - (8)  20.250000 8.211611  6  2.46602036 1.95128006    
#> 65   (7) - (9)  16.250000 8.211611  6  1.97890523 1.90483177    
#> 66   (8) - (9)  -4.000000 8.211611  6 -0.48711513 0.64346045    
stopCluster(cl)

cl <- makeCluster(getOption("cl.cores", ncores))
pout2 <- pairwise.augmentedRCBD(out2, cl = cl,
                                p.adjust = "none")
pout2
#>       contrast   estimate       SE df     t.ratio    p.value sig
#> 1     1 - (10)   7.416667 6.704752  6  1.10618056 1.68898284    
#> 2     1 - (11)  -1.833333 6.704752  6 -0.27343789 0.79368531    
#> 3     1 - (12)   5.166667 6.704752  6  0.77059769 1.52980915    
#> 4        1 - 2   5.666667 4.240458  6  1.33633373 1.77011499    
#> 5        1 - 3   2.666667 4.240458  6  0.62886293 1.44738825    
#> 6        1 - 4   1.333333 4.240458  6  0.31443147 1.23616021    
#> 7      1 - (5)   6.416667 6.704752  6  0.95703262 1.62449310    
#> 8      1 - (6)   6.416667 6.704752  6  0.95703262 1.62449310    
#> 9      1 - (7)  -8.833333 6.704752  6 -1.31747347 0.23575127    
#> 10     1 - (8)  11.416667 6.704752  6  1.70277232 1.86049532    
#> 11     1 - (9)   7.416667 6.704752  6  1.10618056 1.68898284    
#> 12 (10) - (11)  -9.250000 8.211611  6 -1.12645375 0.30300033    
#> 13 (10) - (12)  -2.250000 8.211611  6 -0.27400226 0.79327167    
#> 14    (10) - 2  -1.750000 6.704752  6 -0.26100890 0.80281330    
#> 15    (10) - 3  -4.750000 6.704752  6 -0.70845272 0.50524328    
#> 16    (10) - 4  -6.083333 6.704752  6 -0.90731664 0.39921196    
#> 17  (10) - (5)  -1.000000 8.211611  6 -0.12177878 0.90705045    
#> 18  (10) - (6)  -1.000000 7.344688  6 -0.13615282 0.89615381    
#> 19  (10) - (7) -16.250000 8.211611  6 -1.97890523 0.09516823    
#> 20  (10) - (8)   4.000000 7.344688  6  0.54461128 1.39434533    
#> 21  (10) - (9)   0.000000 8.211611  6  0.00000000 1.00000000    
#> 22 (11) - (12)   7.000000 7.344688  6  0.95306973 1.62264457    
#> 23    (11) - 2   7.500000 6.704752  6  1.11860955 1.69391854    
#> 24    (11) - 3   4.500000 6.704752  6  0.67116573 1.47290592    
#> 25    (11) - 4   3.166667 6.704752  6  0.47230181 1.34659999    
#> 26  (11) - (5)   8.250000 8.211611  6  1.00467496 1.64616381    
#> 27  (11) - (6)   8.250000 8.211611  6  1.00467496 1.64616381    
#> 28  (11) - (7)  -7.000000 7.344688  6 -0.95306973 0.37735543    
#> 29  (11) - (8)  13.250000 8.211611  6  1.61356888 1.84225214    
#> 30  (11) - (9)   9.250000 8.211611  6  1.12645375 1.69699967    
#> 31    (12) - 2   0.500000 6.704752  6  0.07457397 1.05702215    
#> 32    (12) - 3  -2.500000 6.704752  6 -0.37286985 0.72206273    
#> 33    (12) - 4  -3.833333 6.704752  6 -0.57173377 0.58826470    
#> 34  (12) - (5)   1.250000 8.211611  6  0.15222348 1.11599954    
#> 35  (12) - (6)   1.250000 8.211611  6  0.15222348 1.11599954    
#> 36  (12) - (7) -14.000000 7.344688  6 -1.90613947 0.10526990    
#> 37  (12) - (8)   6.250000 8.211611  6  0.76111740 1.52457343    
#> 38  (12) - (9)   2.250000 8.211611  6  0.27400226 1.20672833    
#> 39       2 - 3  -3.000000 4.240458  6 -0.70747080 0.50581089    
#> 40       2 - 4  -4.333333 4.240458  6 -1.02190227 0.34624985    
#> 41     2 - (5)   0.750000 6.704752  6  0.11186096 1.08541796    
#> 42     2 - (6)   0.750000 6.704752  6  0.11186096 1.08541796    
#> 43     2 - (7) -14.500000 6.704752  6 -2.16264514 0.07380701    
#> 44     2 - (8)   5.750000 6.704752  6  0.85760066 1.57596055    
#> 45     2 - (9)   1.750000 6.704752  6  0.26100890 1.19718670    
#> 46       3 - 4  -1.333333 4.240458  6 -0.31443147 0.76383979    
#> 47     3 - (5)   3.750000 6.704752  6  0.55930478 1.40380253    
#> 48     3 - (6)   3.750000 6.704752  6  0.55930478 1.40380253    
#> 49     3 - (7) -11.500000 6.704752  6 -1.71520132 0.13713052    
#> 50     3 - (8)   8.750000 6.704752  6  1.30504448 1.76031005    
#> 51     3 - (9)   4.750000 6.704752  6  0.70845272 1.49475672    
#> 52     4 - (5)   5.083333 6.704752  6  0.75816870 1.52293671    
#> 53     4 - (6)   5.083333 6.704752  6  0.75816870 1.52293671    
#> 54     4 - (7) -10.166667 6.704752  6 -1.51633739 0.18022066    
#> 55     4 - (8)  10.083333 6.704752  6  1.50390840 1.81669827    
#> 56     4 - (9)   6.083333 6.704752  6  0.90731664 1.60078804    
#> 57   (5) - (6)   0.000000 8.211611  6  0.00000000 1.00000000    
#> 58   (5) - (7) -15.250000 8.211611  6 -1.85712645 0.11267166    
#> 59   (5) - (8)   5.000000 8.211611  6  0.60889392 1.43507937    
#> 60   (5) - (9)   1.000000 7.344688  6  0.13615282 1.10384619    
#> 61   (6) - (7) -15.250000 8.211611  6 -1.85712645 0.11267166    
#> 62   (6) - (8)   5.000000 7.344688  6  0.68076409 1.47858869    
#> 63   (6) - (9)   1.000000 8.211611  6  0.12177878 1.09294955    
#> 64   (7) - (8)  20.250000 8.211611  6  2.46602036 1.95128006    
#> 65   (7) - (9)  16.250000 8.211611  6  1.97890523 1.90483177    
#> 66   (8) - (9)  -4.000000 8.211611  6 -0.48711513 0.64346045    
stopCluster(cl)

# Pairwise t test with tukey adjustment ----
cl <- makeCluster(getOption("cl.cores", ncores))
pout1_tukey <- pairwise.augmentedRCBD(out1, cl = cl,
                                      p.adjust = "tukey")
pout1_tukey
#>       contrast   estimate       SE df     t.ratio   p.value sig
#> 1     1 - (10)   7.416667 6.704752  6  1.10618056 0.9985579    
#> 2     1 - (11)  -1.833333 6.704752  6 -0.27343789 1.0000000    
#> 3     1 - (12)   5.166667 6.704752  6  0.77059769 0.9999418    
#> 4        1 - 2   5.666667 4.240458  6  1.33633373 0.9937163    
#> 5        1 - 3   2.666667 4.240458  6  0.62886293 0.9999919    
#> 6        1 - 4   1.333333 4.240458  6  0.31443147 1.0000000    
#> 7      1 - (5)   6.416667 6.704752  6  0.95703262 0.9995791    
#> 8      1 - (6)   6.416667 6.704752  6  0.95703262 0.9995791    
#> 9      1 - (7)  -8.833333 6.704752  6 -1.31747347 1.0000000    
#> 10     1 - (8)  11.416667 6.704752  6  1.70277232 0.9682124    
#> 11     1 - (9)   7.416667 6.704752  6  1.10618056 0.9985579    
#> 12 (10) - (11)  -9.250000 8.211611  6 -1.12645375 1.0000000    
#> 13 (10) - (12)  -2.250000 8.211611  6 -0.27400226 1.0000000    
#> 14    (10) - 2  -1.750000 6.704752  6 -0.26100890 1.0000000    
#> 15    (10) - 3  -4.750000 6.704752  6 -0.70845272 1.0000000    
#> 16    (10) - 4  -6.083333 6.704752  6 -0.90731664 1.0000000    
#> 17  (10) - (5)  -1.000000 8.211611  6 -0.12177878 1.0000000    
#> 18  (10) - (6)  -1.000000 7.344688  6 -0.13615282 1.0000000    
#> 19  (10) - (7) -16.250000 8.211611  6 -1.97890523 1.0000000    
#> 20  (10) - (8)   4.000000 7.344688  6  0.54461128 0.9999981    
#> 21  (10) - (9)   0.000000 8.211611  6  0.00000000 1.0000000    
#> 22 (11) - (12)   7.000000 7.344688  6  0.95306973 0.9995942    
#> 23    (11) - 2   7.500000 6.704752  6  1.11860955 0.9984198    
#> 24    (11) - 3   4.500000 6.704752  6  0.67116573 0.9999846    
#> 25    (11) - 4   3.166667 6.704752  6  0.47230181 0.9999996    
#> 26  (11) - (5)   8.250000 8.211611  6  1.00467496 0.9993581    
#> 27  (11) - (6)   8.250000 8.211611  6  1.00467496 0.9993581    
#> 28  (11) - (7)  -7.000000 7.344688  6 -0.95306973 1.0000000    
#> 29  (11) - (8)  13.250000 8.211611  6  1.61356888 0.9771900    
#> 30  (11) - (9)   9.250000 8.211611  6  1.12645375 0.9983271    
#> 31    (12) - 2   0.500000 6.704752  6  0.07457397 1.0000000    
#> 32    (12) - 3  -2.500000 6.704752  6 -0.37286985 1.0000000    
#> 33    (12) - 4  -3.833333 6.704752  6 -0.57173377 1.0000000    
#> 34  (12) - (5)   1.250000 8.211611  6  0.15222348 1.0000000    
#> 35  (12) - (6)   1.250000 8.211611  6  0.15222348 1.0000000    
#> 36  (12) - (7) -14.000000 7.344688  6 -1.90613947 1.0000000    
#> 37  (12) - (8)   6.250000 8.211611  6  0.76111740 0.9999482    
#> 38  (12) - (9)   2.250000 8.211611  6  0.27400226 1.0000000    
#> 39       2 - 3  -3.000000 4.240458  6 -0.70747080 1.0000000    
#> 40       2 - 4  -4.333333 4.240458  6 -1.02190227 1.0000000    
#> 41     2 - (5)   0.750000 6.704752  6  0.11186096 1.0000000    
#> 42     2 - (6)   0.750000 6.704752  6  0.11186096 1.0000000    
#> 43     2 - (7) -14.500000 6.704752  6 -2.16264514 1.0000000    
#> 44     2 - (8)   5.750000 6.704752  6  0.85760066 0.9998426    
#> 45     2 - (9)   1.750000 6.704752  6  0.26100890 1.0000000    
#> 46       3 - 4  -1.333333 4.240458  6 -0.31443147 1.0000000    
#> 47     3 - (5)   3.750000 6.704752  6  0.55930478 0.9999975    
#> 48     3 - (6)   3.750000 6.704752  6  0.55930478 0.9999975    
#> 49     3 - (7) -11.500000 6.704752  6 -1.71520132 1.0000000    
#> 50     3 - (8)   8.750000 6.704752  6  1.30504448 0.9947262    
#> 51     3 - (9)   4.750000 6.704752  6  0.70845272 0.9999740    
#> 52     4 - (5)   5.083333 6.704752  6  0.75816870 0.9999501    
#> 53     4 - (6)   5.083333 6.704752  6  0.75816870 0.9999501    
#> 54     4 - (7) -10.166667 6.704752  6 -1.51633739 1.0000000    
#> 55     4 - (8)  10.083333 6.704752  6  1.50390840 0.9855836    
#> 56     4 - (9)   6.083333 6.704752  6  0.90731664 0.9997378    
#> 57   (5) - (6)   0.000000 8.211611  6  0.00000000 1.0000000    
#> 58   (5) - (7) -15.250000 8.211611  6 -1.85712645 1.0000000    
#> 59   (5) - (8)   5.000000 8.211611  6  0.60889392 0.9999941    
#> 60   (5) - (9)   1.000000 7.344688  6  0.13615282 1.0000000    
#> 61   (6) - (7) -15.250000 8.211611  6 -1.85712645 1.0000000    
#> 62   (6) - (8)   5.000000 7.344688  6  0.68076409 0.9999823    
#> 63   (6) - (9)   1.000000 8.211611  6  0.12177878 1.0000000    
#> 64   (7) - (8)  20.250000 8.211611  6  2.46602036 0.8043427    
#> 65   (7) - (9)  16.250000 8.211611  6  1.97890523 0.9265988    
#> 66   (8) - (9)  -4.000000 8.211611  6 -0.48711513 1.0000000    
stopCluster(cl)

cl <- makeCluster(getOption("cl.cores", ncores))
pout2_tukey <- pairwise.augmentedRCBD(out2, cl = cl,
                                      p.adjust = "tukey")
pout2_tukey
#>       contrast   estimate       SE df     t.ratio   p.value sig
#> 1     1 - (10)   7.416667 6.704752  6  1.10618056 0.9985579    
#> 2     1 - (11)  -1.833333 6.704752  6 -0.27343789 1.0000000    
#> 3     1 - (12)   5.166667 6.704752  6  0.77059769 0.9999418    
#> 4        1 - 2   5.666667 4.240458  6  1.33633373 0.9937163    
#> 5        1 - 3   2.666667 4.240458  6  0.62886293 0.9999919    
#> 6        1 - 4   1.333333 4.240458  6  0.31443147 1.0000000    
#> 7      1 - (5)   6.416667 6.704752  6  0.95703262 0.9995791    
#> 8      1 - (6)   6.416667 6.704752  6  0.95703262 0.9995791    
#> 9      1 - (7)  -8.833333 6.704752  6 -1.31747347 1.0000000    
#> 10     1 - (8)  11.416667 6.704752  6  1.70277232 0.9682124    
#> 11     1 - (9)   7.416667 6.704752  6  1.10618056 0.9985579    
#> 12 (10) - (11)  -9.250000 8.211611  6 -1.12645375 1.0000000    
#> 13 (10) - (12)  -2.250000 8.211611  6 -0.27400226 1.0000000    
#> 14    (10) - 2  -1.750000 6.704752  6 -0.26100890 1.0000000    
#> 15    (10) - 3  -4.750000 6.704752  6 -0.70845272 1.0000000    
#> 16    (10) - 4  -6.083333 6.704752  6 -0.90731664 1.0000000    
#> 17  (10) - (5)  -1.000000 8.211611  6 -0.12177878 1.0000000    
#> 18  (10) - (6)  -1.000000 7.344688  6 -0.13615282 1.0000000    
#> 19  (10) - (7) -16.250000 8.211611  6 -1.97890523 1.0000000    
#> 20  (10) - (8)   4.000000 7.344688  6  0.54461128 0.9999981    
#> 21  (10) - (9)   0.000000 8.211611  6  0.00000000 1.0000000    
#> 22 (11) - (12)   7.000000 7.344688  6  0.95306973 0.9995942    
#> 23    (11) - 2   7.500000 6.704752  6  1.11860955 0.9984198    
#> 24    (11) - 3   4.500000 6.704752  6  0.67116573 0.9999846    
#> 25    (11) - 4   3.166667 6.704752  6  0.47230181 0.9999996    
#> 26  (11) - (5)   8.250000 8.211611  6  1.00467496 0.9993581    
#> 27  (11) - (6)   8.250000 8.211611  6  1.00467496 0.9993581    
#> 28  (11) - (7)  -7.000000 7.344688  6 -0.95306973 1.0000000    
#> 29  (11) - (8)  13.250000 8.211611  6  1.61356888 0.9771900    
#> 30  (11) - (9)   9.250000 8.211611  6  1.12645375 0.9983271    
#> 31    (12) - 2   0.500000 6.704752  6  0.07457397 1.0000000    
#> 32    (12) - 3  -2.500000 6.704752  6 -0.37286985 1.0000000    
#> 33    (12) - 4  -3.833333 6.704752  6 -0.57173377 1.0000000    
#> 34  (12) - (5)   1.250000 8.211611  6  0.15222348 1.0000000    
#> 35  (12) - (6)   1.250000 8.211611  6  0.15222348 1.0000000    
#> 36  (12) - (7) -14.000000 7.344688  6 -1.90613947 1.0000000    
#> 37  (12) - (8)   6.250000 8.211611  6  0.76111740 0.9999482    
#> 38  (12) - (9)   2.250000 8.211611  6  0.27400226 1.0000000    
#> 39       2 - 3  -3.000000 4.240458  6 -0.70747080 1.0000000    
#> 40       2 - 4  -4.333333 4.240458  6 -1.02190227 1.0000000    
#> 41     2 - (5)   0.750000 6.704752  6  0.11186096 1.0000000    
#> 42     2 - (6)   0.750000 6.704752  6  0.11186096 1.0000000    
#> 43     2 - (7) -14.500000 6.704752  6 -2.16264514 1.0000000    
#> 44     2 - (8)   5.750000 6.704752  6  0.85760066 0.9998426    
#> 45     2 - (9)   1.750000 6.704752  6  0.26100890 1.0000000    
#> 46       3 - 4  -1.333333 4.240458  6 -0.31443147 1.0000000    
#> 47     3 - (5)   3.750000 6.704752  6  0.55930478 0.9999975    
#> 48     3 - (6)   3.750000 6.704752  6  0.55930478 0.9999975    
#> 49     3 - (7) -11.500000 6.704752  6 -1.71520132 1.0000000    
#> 50     3 - (8)   8.750000 6.704752  6  1.30504448 0.9947262    
#> 51     3 - (9)   4.750000 6.704752  6  0.70845272 0.9999740    
#> 52     4 - (5)   5.083333 6.704752  6  0.75816870 0.9999501    
#> 53     4 - (6)   5.083333 6.704752  6  0.75816870 0.9999501    
#> 54     4 - (7) -10.166667 6.704752  6 -1.51633739 1.0000000    
#> 55     4 - (8)  10.083333 6.704752  6  1.50390840 0.9855836    
#> 56     4 - (9)   6.083333 6.704752  6  0.90731664 0.9997378    
#> 57   (5) - (6)   0.000000 8.211611  6  0.00000000 1.0000000    
#> 58   (5) - (7) -15.250000 8.211611  6 -1.85712645 1.0000000    
#> 59   (5) - (8)   5.000000 8.211611  6  0.60889392 0.9999941    
#> 60   (5) - (9)   1.000000 7.344688  6  0.13615282 1.0000000    
#> 61   (6) - (7) -15.250000 8.211611  6 -1.85712645 1.0000000    
#> 62   (6) - (8)   5.000000 7.344688  6  0.68076409 0.9999823    
#> 63   (6) - (9)   1.000000 8.211611  6  0.12177878 1.0000000    
#> 64   (7) - (8)  20.250000 8.211611  6  2.46602036 0.8043427    
#> 65   (7) - (9)  16.250000 8.211611  6  1.97890523 0.9265988    
#> 66   (8) - (9)  -4.000000 8.211611  6 -0.48711513 1.0000000    
stopCluster(cl)

# Pairwise t test with sidak p value adjustment ----
cl <- makeCluster(getOption("cl.cores", ncores))
pout1_sidak <- pairwise.augmentedRCBD(out1, cl = cl,
                                      p.adjust = "sidak")
pout1_sidak
#>       contrast   estimate       SE df     t.ratio   p.value sig
#> 1     1 - (10)   7.416667 6.704752  6  1.10618056 1.0000000    
#> 2     1 - (11)  -1.833333 6.704752  6 -0.27343789 1.0000000    
#> 3     1 - (12)   5.166667 6.704752  6  0.77059769 1.0000000    
#> 4        1 - 2   5.666667 4.240458  6  1.33633373 1.0000000    
#> 5        1 - 3   2.666667 4.240458  6  0.62886293 1.0000000    
#> 6        1 - 4   1.333333 4.240458  6  0.31443147 1.0000000    
#> 7      1 - (5)   6.416667 6.704752  6  0.95703262 1.0000000    
#> 8      1 - (6)   6.416667 6.704752  6  0.95703262 1.0000000    
#> 9      1 - (7)  -8.833333 6.704752  6 -1.31747347 1.0000000    
#> 10     1 - (8)  11.416667 6.704752  6  1.70277232 0.9999506    
#> 11     1 - (9)   7.416667 6.704752  6  1.10618056 1.0000000    
#> 12 (10) - (11)  -9.250000 8.211611  6 -1.12645375 1.0000000    
#> 13 (10) - (12)  -2.250000 8.211611  6 -0.27400226 1.0000000    
#> 14    (10) - 2  -1.750000 6.704752  6 -0.26100890 1.0000000    
#> 15    (10) - 3  -4.750000 6.704752  6 -0.70845272 1.0000000    
#> 16    (10) - 4  -6.083333 6.704752  6 -0.90731664 1.0000000    
#> 17  (10) - (5)  -1.000000 8.211611  6 -0.12177878 1.0000000    
#> 18  (10) - (6)  -1.000000 7.344688  6 -0.13615282 1.0000000    
#> 19  (10) - (7) -16.250000 8.211611  6 -1.97890523 0.9986402    
#> 20  (10) - (8)   4.000000 7.344688  6  0.54461128 1.0000000    
#> 21  (10) - (9)   0.000000 8.211611  6  0.00000000 1.0000000    
#> 22 (11) - (12)   7.000000 7.344688  6  0.95306973 1.0000000    
#> 23    (11) - 2   7.500000 6.704752  6  1.11860955 1.0000000    
#> 24    (11) - 3   4.500000 6.704752  6  0.67116573 1.0000000    
#> 25    (11) - 4   3.166667 6.704752  6  0.47230181 1.0000000    
#> 26  (11) - (5)   8.250000 8.211611  6  1.00467496 1.0000000    
#> 27  (11) - (6)   8.250000 8.211611  6  1.00467496 1.0000000    
#> 28  (11) - (7)  -7.000000 7.344688  6 -0.95306973 1.0000000    
#> 29  (11) - (8)  13.250000 8.211611  6  1.61356888 0.9999880    
#> 30  (11) - (9)   9.250000 8.211611  6  1.12645375 1.0000000    
#> 31    (12) - 2   0.500000 6.704752  6  0.07457397 1.0000000    
#> 32    (12) - 3  -2.500000 6.704752  6 -0.37286985 1.0000000    
#> 33    (12) - 4  -3.833333 6.704752  6 -0.57173377 1.0000000    
#> 34  (12) - (5)   1.250000 8.211611  6  0.15222348 1.0000000    
#> 35  (12) - (6)   1.250000 8.211611  6  0.15222348 1.0000000    
#> 36  (12) - (7) -14.000000 7.344688  6 -1.90613947 0.9993519    
#> 37  (12) - (8)   6.250000 8.211611  6  0.76111740 1.0000000    
#> 38  (12) - (9)   2.250000 8.211611  6  0.27400226 1.0000000    
#> 39       2 - 3  -3.000000 4.240458  6 -0.70747080 1.0000000    
#> 40       2 - 4  -4.333333 4.240458  6 -1.02190227 1.0000000    
#> 41     2 - (5)   0.750000 6.704752  6  0.11186096 1.0000000    
#> 42     2 - (6)   0.750000 6.704752  6  0.11186096 1.0000000    
#> 43     2 - (7) -14.500000 6.704752  6 -2.16264514 0.9936569    
#> 44     2 - (8)   5.750000 6.704752  6  0.85760066 1.0000000    
#> 45     2 - (9)   1.750000 6.704752  6  0.26100890 1.0000000    
#> 46       3 - 4  -1.333333 4.240458  6 -0.31443147 1.0000000    
#> 47     3 - (5)   3.750000 6.704752  6  0.55930478 1.0000000    
#> 48     3 - (6)   3.750000 6.704752  6  0.55930478 1.0000000    
#> 49     3 - (7) -11.500000 6.704752  6 -1.71520132 0.9999408    
#> 50     3 - (8)   8.750000 6.704752  6  1.30504448 1.0000000    
#> 51     3 - (9)   4.750000 6.704752  6  0.70845272 1.0000000    
#> 52     4 - (5)   5.083333 6.704752  6  0.75816870 1.0000000    
#> 53     4 - (6)   5.083333 6.704752  6  0.75816870 1.0000000    
#> 54     4 - (7) -10.166667 6.704752  6 -1.51633739 0.9999980    
#> 55     4 - (8)  10.083333 6.704752  6  1.50390840 0.9999984    
#> 56     4 - (9)   6.083333 6.704752  6  0.90731664 1.0000000    
#> 57   (5) - (6)   0.000000 8.211611  6  0.00000000 1.0000000    
#> 58   (5) - (7) -15.250000 8.211611  6 -1.85712645 0.9996254    
#> 59   (5) - (8)   5.000000 8.211611  6  0.60889392 1.0000000    
#> 60   (5) - (9)   1.000000 7.344688  6  0.13615282 1.0000000    
#> 61   (6) - (7) -15.250000 8.211611  6 -1.85712645 0.9996254    
#> 62   (6) - (8)   5.000000 7.344688  6  0.68076409 1.0000000    
#> 63   (6) - (9)   1.000000 8.211611  6  0.12177878 1.0000000    
#> 64   (7) - (8)  20.250000 8.211611  6  2.46602036 0.9629870    
#> 65   (7) - (9)  16.250000 8.211611  6  1.97890523 0.9986402    
#> 66   (8) - (9)  -4.000000 8.211611  6 -0.48711513 1.0000000    
stopCluster(cl)

cl <- makeCluster(getOption("cl.cores", ncores))
pout2_sidak <- pairwise.augmentedRCBD(out2, cl = cl,
                                      p.adjust = "sidak")
pout2_sidak
#>       contrast   estimate       SE df     t.ratio   p.value sig
#> 1     1 - (10)   7.416667 6.704752  6  1.10618056 1.0000000    
#> 2     1 - (11)  -1.833333 6.704752  6 -0.27343789 1.0000000    
#> 3     1 - (12)   5.166667 6.704752  6  0.77059769 1.0000000    
#> 4        1 - 2   5.666667 4.240458  6  1.33633373 1.0000000    
#> 5        1 - 3   2.666667 4.240458  6  0.62886293 1.0000000    
#> 6        1 - 4   1.333333 4.240458  6  0.31443147 1.0000000    
#> 7      1 - (5)   6.416667 6.704752  6  0.95703262 1.0000000    
#> 8      1 - (6)   6.416667 6.704752  6  0.95703262 1.0000000    
#> 9      1 - (7)  -8.833333 6.704752  6 -1.31747347 1.0000000    
#> 10     1 - (8)  11.416667 6.704752  6  1.70277232 0.9999506    
#> 11     1 - (9)   7.416667 6.704752  6  1.10618056 1.0000000    
#> 12 (10) - (11)  -9.250000 8.211611  6 -1.12645375 1.0000000    
#> 13 (10) - (12)  -2.250000 8.211611  6 -0.27400226 1.0000000    
#> 14    (10) - 2  -1.750000 6.704752  6 -0.26100890 1.0000000    
#> 15    (10) - 3  -4.750000 6.704752  6 -0.70845272 1.0000000    
#> 16    (10) - 4  -6.083333 6.704752  6 -0.90731664 1.0000000    
#> 17  (10) - (5)  -1.000000 8.211611  6 -0.12177878 1.0000000    
#> 18  (10) - (6)  -1.000000 7.344688  6 -0.13615282 1.0000000    
#> 19  (10) - (7) -16.250000 8.211611  6 -1.97890523 0.9986402    
#> 20  (10) - (8)   4.000000 7.344688  6  0.54461128 1.0000000    
#> 21  (10) - (9)   0.000000 8.211611  6  0.00000000 1.0000000    
#> 22 (11) - (12)   7.000000 7.344688  6  0.95306973 1.0000000    
#> 23    (11) - 2   7.500000 6.704752  6  1.11860955 1.0000000    
#> 24    (11) - 3   4.500000 6.704752  6  0.67116573 1.0000000    
#> 25    (11) - 4   3.166667 6.704752  6  0.47230181 1.0000000    
#> 26  (11) - (5)   8.250000 8.211611  6  1.00467496 1.0000000    
#> 27  (11) - (6)   8.250000 8.211611  6  1.00467496 1.0000000    
#> 28  (11) - (7)  -7.000000 7.344688  6 -0.95306973 1.0000000    
#> 29  (11) - (8)  13.250000 8.211611  6  1.61356888 0.9999880    
#> 30  (11) - (9)   9.250000 8.211611  6  1.12645375 1.0000000    
#> 31    (12) - 2   0.500000 6.704752  6  0.07457397 1.0000000    
#> 32    (12) - 3  -2.500000 6.704752  6 -0.37286985 1.0000000    
#> 33    (12) - 4  -3.833333 6.704752  6 -0.57173377 1.0000000    
#> 34  (12) - (5)   1.250000 8.211611  6  0.15222348 1.0000000    
#> 35  (12) - (6)   1.250000 8.211611  6  0.15222348 1.0000000    
#> 36  (12) - (7) -14.000000 7.344688  6 -1.90613947 0.9993519    
#> 37  (12) - (8)   6.250000 8.211611  6  0.76111740 1.0000000    
#> 38  (12) - (9)   2.250000 8.211611  6  0.27400226 1.0000000    
#> 39       2 - 3  -3.000000 4.240458  6 -0.70747080 1.0000000    
#> 40       2 - 4  -4.333333 4.240458  6 -1.02190227 1.0000000    
#> 41     2 - (5)   0.750000 6.704752  6  0.11186096 1.0000000    
#> 42     2 - (6)   0.750000 6.704752  6  0.11186096 1.0000000    
#> 43     2 - (7) -14.500000 6.704752  6 -2.16264514 0.9936569    
#> 44     2 - (8)   5.750000 6.704752  6  0.85760066 1.0000000    
#> 45     2 - (9)   1.750000 6.704752  6  0.26100890 1.0000000    
#> 46       3 - 4  -1.333333 4.240458  6 -0.31443147 1.0000000    
#> 47     3 - (5)   3.750000 6.704752  6  0.55930478 1.0000000    
#> 48     3 - (6)   3.750000 6.704752  6  0.55930478 1.0000000    
#> 49     3 - (7) -11.500000 6.704752  6 -1.71520132 0.9999408    
#> 50     3 - (8)   8.750000 6.704752  6  1.30504448 1.0000000    
#> 51     3 - (9)   4.750000 6.704752  6  0.70845272 1.0000000    
#> 52     4 - (5)   5.083333 6.704752  6  0.75816870 1.0000000    
#> 53     4 - (6)   5.083333 6.704752  6  0.75816870 1.0000000    
#> 54     4 - (7) -10.166667 6.704752  6 -1.51633739 0.9999980    
#> 55     4 - (8)  10.083333 6.704752  6  1.50390840 0.9999984    
#> 56     4 - (9)   6.083333 6.704752  6  0.90731664 1.0000000    
#> 57   (5) - (6)   0.000000 8.211611  6  0.00000000 1.0000000    
#> 58   (5) - (7) -15.250000 8.211611  6 -1.85712645 0.9996254    
#> 59   (5) - (8)   5.000000 8.211611  6  0.60889392 1.0000000    
#> 60   (5) - (9)   1.000000 7.344688  6  0.13615282 1.0000000    
#> 61   (6) - (7) -15.250000 8.211611  6 -1.85712645 0.9996254    
#> 62   (6) - (8)   5.000000 7.344688  6  0.68076409 1.0000000    
#> 63   (6) - (9)   1.000000 8.211611  6  0.12177878 1.0000000    
#> 64   (7) - (8)  20.250000 8.211611  6  2.46602036 0.9629870    
#> 65   (7) - (9)  16.250000 8.211611  6  1.97890523 0.9986402    
#> 66   (8) - (9)  -4.000000 8.211611  6 -0.48711513 1.0000000    
stopCluster(cl)